A1A2A3 {VGAM}	R Documentation

The A1A2A3 Blood Group System

Description

Estimates the three independent parameters of the the A1A2A3 blood group system.

Usage

A1A2A3(link = "logit", inbreeding = FALSE, ip1 = NULL, ip2 = NULL, iF = NULL)

Arguments

link
Link function applied to p1, p2 and f. See Links for more choices.

inbreeding
Logical. Is there inbreeding?

ip1, ip2, iF
Optional initial value for p1, p2 and f.

Details

The parameters p1 and p2 are probabilities, so that p3=1-p1-p2 is the third probability. The parameter f is the third independent parameter if inbreeding = TRUE. If inbreeding = FALSE then f = 0 and Hardy-Weinberg Equilibrium (HWE) is assumed.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Note

The input can be a 6-column matrix of counts, with columns corresponding to A1A1, A1A2, A1A3, A2A2, A2A3, A3A3 (in order). Alternatively, the input can be a 6-column matrix of proportions (so each row adds to 1) and the weights argument is used to specify the total number of counts for each row.

Author(s)

T. W. Yee

References

Lange, K. (2002) Mathematical and Statistical Methods for Genetic Analysis, 2nd ed. New York: Springer-Verlag.

See Also

AA.Aa.aa, AB.Ab.aB.ab, ABO, MNSs.

Examples

ymat <- cbind(108, 196, 429, 143, 513, 559)
fit <- vglm(ymat ~ 1, A1A2A3(link = probit), trace = TRUE, crit = "coef")
fit <- vglm(ymat ~ 1, A1A2A3(link = logit, ip1 = 0.3, ip2 = 0.3, iF = 0.02),
            trace = TRUE, crit = "coef")
Coef(fit)  # Estimated p1 and p2
rbind(ymat, sum(ymat) * fitted(fit))
sqrt(diag(vcov(fit)))

[Package VGAM version 1.0-6 Index]