ABO {VGAM}	R Documentation

The ABO Blood Group System

Description

Estimates the two independent parameters of the the ABO blood group system.

Usage

ABO(link.pA = "logit", link.pB = "logit", ipA = NULL, ipB = NULL,
    ipO = NULL, zero = NULL)

Arguments

link.pA, link.pB
Link functions applied to pA and pB. See Links for more choices.

ipA, ipB, ipO
Optional initial value for pA and pB and pO. A NULL value means values are computed internally.

zero
Details at CommonVGAMffArguments.

Details

The parameters pA and pB are probabilities, so that pO=1-pA-pB is the third probability. The probabilities pA and pB correspond to A and B respectively, so that pO is the probability for O. It is easier to make use of initial values for pO than for pB. In documentation elsewhere I sometimes use pA=p, pB=q, pO=r.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Note

The input can be a 4-column matrix of counts, where the columns are A, B, AB, O (in order). Alternatively, the input can be a 4-column matrix of proportions (so each row adds to 1) and the weights argument is used to specify the total number of counts for each row.

Author(s)

T. W. Yee

References

Lange, K. (2002) Mathematical and Statistical Methods for Genetic Analysis, 2nd ed. New York: Springer-Verlag.

See Also

AA.Aa.aa, AB.Ab.aB.ab, A1A2A3, MNSs.

Examples

ymat <- cbind(A = 725, B = 258, AB = 72, O = 1073)  # Order matters, not the name
fit <- vglm(ymat ~ 1, ABO(link.pA = "identitylink",
                          link.pB = "identitylink"), trace = TRUE,
            crit = "coef")
coef(fit, matrix = TRUE)
Coef(fit)  # Estimated pA and pB
rbind(ymat, sum(ymat) * fitted(fit))
sqrt(diag(vcov(fit)))

[Package VGAM version 1.0-6 Index]