Coef {VGAM}	R Documentation

Computes Model Coefficients and Quantities

Description

Coef is a generic function which computes model coefficients from objects returned by modelling functions. It is an auxiliary function to coef that enables extra capabilities for some specific models.

Usage

Coef(object, ...)

Arguments

object
An object for which the computation of other types of model coefficients or quantities is meaningful.

...
Other arguments fed into the specific methods function of the model.

Details

This function can often be useful for vglm objects with just an intercept term in the RHS of the formula, e.g., y ~ 1. Then often this function will apply the inverse link functions to the parameters. See the example below.

For reduced-rank VGLMs, this function can return the A, C matrices, etc.

For quadratic and additive ordination models, this function can return ecological meaningful quantities such as tolerances, optimums, maximums.

Value

The value returned depends specifically on the methods function invoked.

Warning

This function may not work for all VGAM family functions. You should check your results on some artificial data before applying it to models fitted to real data.

Author(s)

Thomas W. Yee

References

Yee, T. W. and Hastie, T. J. (2003) Reduced-rank vector generalized linear models. Statistical Modelling, 3, 15–41.

See Also

coef, Coef.vlm, Coef.rrvglm, Coef.qrrvglm, depvar.

Examples

nn <- 1000
bdata <- data.frame(y = rbeta(nn, shape1 = 1, shape2 = 3))  # Original scale
fit <- vglm(y ~ 1, betaR, data = bdata, trace = TRUE)  # Intercept-only model
coef(fit, matrix = TRUE)  # Both on a log scale
Coef(fit)  # On the original scale

[Package VGAM version 1.0-6 Index]