Coef.qrrvglm {VGAM}	R Documentation

Returns Important Matrices etc. of a QO Object

Description

This methods function returns important matrices etc. of a QO object.

Usage

Coef.qrrvglm(object, varI.latvar = FALSE, refResponse = NULL, ...)

Arguments

object
A CQO object. The former has class "qrrvglm".

varI.latvar
Logical indicating whether to scale the site scores (latent variables) to have variance-covariance matrix equal to the rank-R identity matrix. All models have uncorrelated site scores (latent variables), and this option stretches or shrinks the ordination axes if TRUE. See below for further details.

refResponse
Integer or character. Specifies the reference response or reference species. By default, the reference species is found by searching sequentially starting from the first species until a positive-definite tolerance matrix is found. Then this tolerance matrix is transformed to the identity matrix. Then the sites scores (latent variables) are made uncorrelated. See below for further details.

...
Currently unused.

Details

If I.tolerances=TRUE or eq.tolerances=TRUE (and its estimated tolerance matrix is positive-definite) then all species' tolerances are unity by transformation or by definition, and the spread of the site scores can be compared to them. Vice versa, if one wishes to compare the tolerances with the sites score variability then setting varI.latvar=TRUE is more appropriate.

For rank-2 QRR-VGLMs, one of the species can be chosen so that the angle of its major axis and minor axis is zero, i.e., parallel to the ordination axes. This means the effect on the latent vars is independent on that species, and that its tolerance matrix is diagonal. The argument refResponse allows one to choose which is the reference species, which must have a positive-definite tolerance matrix, i.e., is bell-shaped. If refResponse is not specified, then the code will try to choose some reference species starting from the first species. Although the refResponse argument could possibly be offered as an option when fitting the model, it is currently available after fitting the model, e.g., in the functions Coef.qrrvglm and lvplot.qrrvglm.

Value

The A, B1, C, T, D matrices/arrays are returned, along with other slots. The returned object has class "Coef.qrrvglm" (see Coef.qrrvglm-class).

Note

Consider an equal-tolerances Poisson/binomial CQO model with noRRR = ~ 1. For R=1 it has about 2*S+p2 parameters. For R=2 it has about 3*S+2*p_2 parameters. Here, S is the number of species, and p2=p-1 is the number of environmental variables making up the latent variable. For an unequal-tolerances Poisson/binomial CQO model with noRRR = ~ 1, it has about 3*S-1+p2 parameters for R=1, and about 6*S -3 +2*p2 parameters for R=2. Since the total number of data points is n*S, where n is the number of sites, it pays to divide the number of data points by the number of parameters to get some idea about how much information the parameters contain.

Author(s)

Thomas W. Yee

References

Yee, T. W. (2004) A new technique for maximum-likelihood canonical Gaussian ordination. Ecological Monographs, 74, 685–701.

Yee, T. W. (2006) Constrained additive ordination. Ecology, 87, 203–213.

See Also

cqo, Coef.qrrvglm-class, print.Coef.qrrvglm, lvplot.qrrvglm.

Examples

set.seed(123)
x2 <- rnorm(n <- 100)
x3 <- rnorm(n)
x4 <- rnorm(n)
latvar1 <- 0 + x3 - 2*x4
lambda1 <- exp(3 - 0.5 * ( latvar1-0)^2)
lambda2 <- exp(2 - 0.5 * ( latvar1-1)^2)
lambda3 <- exp(2 - 0.5 * ((latvar1+4)/2)^2)  # Unequal tolerances
y1 <- rpois(n, lambda1)
y2 <- rpois(n, lambda2)
y3 <- rpois(n, lambda3)
set.seed(111)
# vvv p1 <- cqo(cbind(y1, y2, y3) ~ x2 + x3 + x4, poissonff, trace = FALSE)
## Not run:  lvplot(p1, y = TRUE, lcol = 1:3, pch = 1:3, pcol = 1:3)

# vvv Coef(p1)
# vvv print(Coef(p1), digits=3)

[Package VGAM version 1.0-6 Index]