Coef.qrrvglm-class {VGAM}	R Documentation

Class “Coef.qrrvglm”

Description

The most pertinent matrices and other quantities pertaining to a QRR-VGLM (CQO model).

Objects from the Class

Objects can be created by calls of the form Coef(object, ...) where object is an object of class "qrrvglm" (created by cqo).

In this document, R is the rank, M is the number of linear predictors and n is the number of observations.

Slots

A:

Of class "matrix", A, which are the linear ‘coefficients’ of the matrix of latent variables. It is M by R.

B1:

Of class "matrix", B1. These correspond to terms of the argument noRRR.

C:

Of class "matrix", C, the canonical coefficients. It has R columns.

Constrained:

Logical. Whether the model is a constrained ordination model.

D:

Of class "array", D[,,j] is an order-Rank matrix, for j = 1,...,M. Ideally, these are negative-definite in order to make the response curves/surfaces bell-shaped.

Rank:

The rank (dimension, number of latent variables) of the RR-VGLM. Called R.

latvar:

n by R matrix of latent variable values.

latvar.order:

Of class "matrix", the permutation returned when the function order is applied to each column of latvar. This enables each column of latvar to be easily sorted.

Maximum:

Of class "numeric", the M maximum fitted values. That is, the fitted values at the optimums for noRRR = ~ 1 models. If noRRR is not ~ 1 then these will be NAs.

NOS:

Number of species.

Optimum:

Of class "matrix", the values of the latent variables where the optimums are. If the curves are not bell-shaped, then the value will be NA or NaN.

Optimum.order:

Of class "matrix", the permutation returned when the function order is applied to each column of Optimum. This enables each row of Optimum to be easily sorted.

bellshaped:

Vector of logicals: is each response curve/surface bell-shaped?

dispersion:

Dispersion parameter(s).

Dzero:

Vector of logicals, is each of the response curves linear in the latent variable(s)? It will be if and only if D[,,j] equals O, for j = 1,...,M .

Tolerance:

Object of class "array", Tolerance[,,j] is an order-Rank matrix, for j = 1,...,M, being the matrix of tolerances (squared if on the diagonal). These are denoted by T in Yee (2004). Ideally, these are positive-definite in order to make the response curves/surfaces bell-shaped. The tolerance matrices satisfy T_s = -(0.5 D_s^(-1).

Author(s)

Thomas W. Yee

References

Yee, T. W. (2004) A new technique for maximum-likelihood canonical Gaussian ordination. Ecological Monographs, 74, 685–701.

See Also

Coef.qrrvglm, cqo, print.Coef.qrrvglm.

Examples

x2 <- rnorm(n <- 100)
x3 <- rnorm(n)
x4 <- rnorm(n)
latvar1 <- 0 + x3 - 2*x4
lambda1 <- exp(3 - 0.5 * ( latvar1-0)^2)
lambda2 <- exp(2 - 0.5 * ( latvar1-1)^2)
lambda3 <- exp(2 - 0.5 * ((latvar1+4)/2)^2)
y1 <- rpois(n, lambda1)
y2 <- rpois(n, lambda2)
y3 <- rpois(n, lambda3)
yy <- cbind(y1, y2, y3)
# vvv p1 <- cqo(yy ~ x2 + x3 + x4, fam = poissonff, trace = FALSE)
## Not run: 
lvplot(p1, y = TRUE, lcol = 1:3, pch = 1:3, pcol = 1:3)

## End(Not run)
# vvv print(Coef(p1), digits = 3)

[Package VGAM version 1.0-6 Index]