Coef.rrvglm-class {VGAM}	R Documentation

Class “Coef.rrvglm”

Description

The most pertinent matrices and other quantities pertaining to a RR-VGLM.

Objects from the Class

Objects can be created by calls of the form Coef(object, ...) where object is an object of class rrvglm (see rrvglm-class).

In this document, M is the number of linear predictors and n is the number of observations.

Slots

A:

Of class "matrix", A.

B1:

Of class "matrix", B1.

C:

Of class "matrix", C.

Rank:

The rank of the RR-VGLM.

colx1.index:

Index of the columns of the "vlm"-type model matrix corresponding to the variables in x1. These correspond to B1.

colx2.index:

Index of the columns of the "vlm"-type model matrix corresponding to the variables in x2. These correspond to the reduced-rank regression.

Atilde:

Object of class "matrix", the A matrix with the corner rows removed. Thus each of the elements have been estimated. This matrix is returned only if corner constraints were used.

Author(s)

Thomas W. Yee

References

Yee, T. W. and Hastie, T. J. (2003) Reduced-rank vector generalized linear models. Statistical Modelling, 3, 15–41.

See Also

Coef.rrvglm, rrvglm, rrvglm-class, print.Coef.rrvglm.

Examples

# Rank-1 stereotype model of Anderson (1984)
pneumo <- transform(pneumo, let = log(exposure.time), x3 = runif(nrow(pneumo)))
fit <- rrvglm(cbind(normal, mild, severe) ~ let + x3, multinomial, data = pneumo)
coef(fit, matrix = TRUE)
Coef(fit)
# print(Coef(fit), digits = 3)

[Package VGAM version 1.0-6 Index]