Coef.vlm {VGAM}	R Documentation

Extract Model Coefficients for VLM Objects

Description

Amongst other things, this function applies inverse link functions to the parameters of intercept-only VGLMs.

Usage

Coef.vlm(object, ...)

Arguments

object
A fitted model.

...
Arguments which may be passed into coef.

Details

Most VGAM family functions apply a link function to the parameters, e.g., positive parameter are often have a log link, parameters between 0 and 1 have a logit link. This function can back-transform the parameter estimate to the original scale.

Value

For intercept-only models (e.g., formula is y ~ 1) the back-transformed parameter estimates can be returned.

Warning

This function may not work for all VGAM family functions. You should check your results on some artificial data before applying it to models fitted to real data.

Author(s)

Thomas W. Yee

References

Yee, T. W. and Hastie, T. J. (2003) Reduced-rank vector generalized linear models. Statistical Modelling, 3, 15–41.

See Also

Coef, coef.

Examples

set.seed(123); nn <- 1000
bdata <- data.frame(y = rbeta(nn, shape1 = 1, shape2 = 3))
fit <- vglm(y ~ 1, betaff, data = bdata, trace = TRUE)  # intercept-only model
coef(fit, matrix = TRUE)  # log scale
Coef(fit)  # On the original scale

[Package VGAM version 1.0-6 Index]