MNSs {VGAM}	R Documentation

The MNSs Blood Group System

Description

Estimates the three independent parameters of the the MNSs blood group system.

Usage

MNSs(link = "logit", imS = NULL, ims = NULL, inS = NULL)

Arguments

link
Link function applied to the three parameters. See Links for more choices.

imS, ims, inS
Optional initial value for mS, ms and nS respectively. A NULL means they are computed internally.

Details

There are three independent parameters: m_S, m_s, n_S, say, so that n_s = 1 - m_S - m_s - n_S. We let the eta vector (transposed) be (g(m_S), g(m_s), g(n_S)) where g is the link function.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Note

The input can be a 6-column matrix of counts, where the columns are MS, Ms, MNS, MNs, NS, Ns (in order). Alternatively, the input can be a 6-column matrix of proportions (so each row adds to 1) and the weights argument is used to specify the total number of counts for each row.

Author(s)

T. W. Yee

References

Elandt-Johnson, R. C. (1971) Probability Models and Statistical Methods in Genetics, New York: Wiley.

See Also

AA.Aa.aa, AB.Ab.aB.ab, ABO, A1A2A3.

Examples

# Order matters only:
y <- cbind(MS = 295, Ms = 107, MNS = 379, MNs = 322, NS = 102, Ns = 214)
fit <- vglm(y ~ 1, MNSs("logit", .25, .28, .08), trace = TRUE)
fit <- vglm(y ~ 1, MNSs(link = logit), trace = TRUE, crit = "coef")
Coef(fit)
rbind(y, sum(y)*fitted(fit))
sqrt(diag(vcov(fit)))

[Package VGAM version 1.0-6 Index]