SurvS4-class {VGAM}	R Documentation

Class "SurvS4"

Description

S4 version of the Surv class.

Objects from the Class

A virtual Class: No objects may be created from it.

Extends

Class "Surv", directly. Class "matrix", directly. Class "oldClass", by class "Surv", distance 2. Class "structure", by class "matrix", distance 2. Class "array", by class "matrix", distance 2. Class "vector", by class "matrix", distance 3, with explicit coerce. Class "vector", by class "matrix", distance 4, with explicit coerce.

Methods

show

signature(object = "SurvS4"): ...

Warning

This code has not been thoroughly tested.

Note

The purpose of having SurvS4 in VGAM is so that the same input can be fed into vglm as functions in survival such as survreg.

Author(s)

T. W. Yee.

References

See survival.

See Also

SurvS4.

Examples

showClass("SurvS4")

[Package VGAM version 1.0-6 Index]