auuc {VGAM}	R Documentation

Auckland University Undergraduate Counts Data

Description

Undergraduate student enrolments at the University of Auckland in 1990.

Usage

data(auuc)

Format

A data frame with 4 observations on the following 5 variables.

Commerce

a numeric vector of counts.

Arts

a numeric vector of counts.

SciEng

a numeric vector of counts.

Law

a numeric vector of counts.

Medicine

a numeric vector of counts.

Details

Each student is cross-classified by their colleges (Science and Engineering have been combined) and the socio-economic status (SES) of their fathers (1 = highest, down to 4 = lowest).

Source

Dr Tony Morrison.

References

Wild, C. J. and Seber, G. A. F. (2000) Chance Encounters: A First Course in Data Analysis and Inference, New York: Wiley.

Examples

auuc
## Not run: 
round(fitted(grc(auuc)))
round(fitted(grc(auuc, Rank = 2)))

## End(Not run)

[Package VGAM version 1.0-6 Index]