bell {VGAM}	R Documentation

The Bell Series of Integers

Description

Returns the values of the Bell series.

Usage

bell(n)

Arguments

n
Vector of non-negative integers. Values greater than 218 return an Inf. Non-integers or negative values return a NaN.

Details

The Bell numbers emerge from a series expansion of exp(exp(x) - 1) for real x. The first few values are B_0 = 1, B_1 = 1, B_2 = 2, B_3 = 5, B_4 = 15. The series increases quickly so that overflow occurs when its argument is more than 218.

Value

This function returns B_n.

Author(s)

T. W. Yee

References

Bell, E. T. (1934) Exponential polynomials. Ann. Math., 35, 258–277.

Bell, E. T. (1934) Exponential numbers. Amer. Math. Monthly, 41, 411–419.

See Also

bellff, rbell.

Examples

 ## Not run: 
plot(0:10, bell(0:10), log = "y", type = "h", las = 1, col = "blue")

## End(Not run)

[Package VGAM version 1.0-6 Index]