Bell {VGAM}	R Documentation

The Bell Distribution

Description

Density, and random generation for the Topp-Leone distribution.

Usage

dbell(x, shape, log = FALSE)
rbell(n, shape)

Arguments

x, n
Same as Uniform.

shape
the (shape) parameter, which is positive.

log
Logical. If log = TRUE then the logarithm of the density is returned.

Details

See bell, the VGAM family function for estimating the parameter s by maximum likelihood estimation.

Value

dbell gives the density, rbell generates random deviates. If shape is large then rbell will become computationally expensive.

See Also

bell.

Examples

## Not run: 
plot(0:15, dbell(0:15, shape = 1.5), type = "h", las = 1, col = "blue")

## End(Not run)

[Package VGAM version 1.0-6 Index]