biamhcop {VGAM}	R Documentation

Ali-Mikhail-Haq Distribution Family Function

Description

Estimate the association parameter of Ali-Mikhail-Haq's bivariate distribution by maximum likelihood estimation.

Usage

biamhcop(lapar = "rhobit", iapar = NULL, imethod = 1, nsimEIM = 250)

Arguments

lapar
Link function applied to the association parameter alpha, which is real and -1 < alpha < 1. See Links for more choices.

iapar
Numeric. Optional initial value for alpha. By default, an initial value is chosen internally. If a convergence failure occurs try assigning a different value. Assigning a value will override the argument imethod.

imethod
An integer with value 1 or 2 which specifies the initialization method. If failure to converge occurs try the other value, or else specify a value for iapar.

nsimEIM
See CommonVGAMffArguments for more information.

Details

The cumulative distribution function is

P(Y1 < = y1, Y2 < = y2) = y1 * y2 / ( 1 - alpha * (1 - y1) * (1 - y2) )

for -1 < alpha < 1. The support of the function is the unit square. The marginal distributions are the standard uniform distributions. When alpha = 0 the random variables are independent. This is an Archimedean copula.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Note

The response must be a two-column matrix. Currently, the fitted value is a matrix with two columns and values equal to 0.5. This is because each marginal distribution corresponds to a standard uniform distribution.

Author(s)

T. W. Yee and C. S. Chee

References

Balakrishnan, N. and Lai, C.-D. (2009) Continuous Bivariate Distributions, 2nd ed. New York: Springer.

See Also

rbiamhcop, bifgmcop, bigumbelIexp, rbilogis, simulate.vlm.

Examples

ymat <- rbiamhcop(1000, apar = rhobit(2, inverse = TRUE))
fit <- vglm(ymat ~ 1, biamhcop, trace = TRUE)
coef(fit, matrix = TRUE)
Coef(fit)

[Package VGAM version 1.0-6 Index]