bifgmcop {VGAM}	R Documentation

Farlie-Gumbel-Morgenstern's Bivariate Distribution Family Function

Description

Estimate the association parameter of Farlie-Gumbel-Morgenstern's bivariate distribution by maximum likelihood estimation.

Usage

bifgmcop(lapar = "rhobit", iapar = NULL, imethod = 1)

Arguments

lapar, iapar, imethod
Details at CommonVGAMffArguments. See Links for more link function choices.

Details

The cumulative distribution function is

P(Y1 <= y1, Y2 <= y2) = y1 * y2 * ( 1 + alpha * (1 - y1) * (1 - y2) )

for -1 < alpha < 1. The support of the function is the unit square. The marginal distributions are the standard uniform distributions. When alpha=0 the random variables are independent.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Note

The response must be a two-column matrix. Currently, the fitted value is a matrix with two columns and values equal to 0.5. This is because each marginal distribution corresponds to a standard uniform distribution.

Author(s)

T. W. Yee

References

Castillo, E., Hadi, A. S., Balakrishnan, N. Sarabia, J. S. (2005) Extreme Value and Related Models with Applications in Engineering and Science, Hoboken, NJ, USA: Wiley-Interscience.

Smith, M. D. (2007) Invariance theorems for Fisher information. Communications in Statistics—Theory and Methods, 36(12), 2213–2222.

See Also

rbifgmcop, bifrankcop, bifgmexp, simulate.vlm.

Examples

ymat <- rbifgmcop(n = 1000, apar = rhobit(3, inverse = TRUE))
## Not run: plot(ymat, col = "blue")
fit <- vglm(ymat ~ 1, fam = bifgmcop, trace = TRUE)
coef(fit, matrix = TRUE)
Coef(fit)
head(fitted(fit))

[Package VGAM version 1.0-6 Index]