Frank {VGAM}	R Documentation

Frank's Bivariate Distribution

Description

Density, distribution function, and random generation for the (one parameter) bivariate Frank distribution.

Usage

dbifrankcop(x1, x2, apar, log = FALSE)
pbifrankcop(q1, q2, apar)
rbifrankcop(n, apar)

Arguments

x1, x2, q1, q2
vector of quantiles.

n
number of observations. Same as in runif.

apar
the positive association parameter.

log
Logical. If log = TRUE then the logarithm of the density is returned.

Details

See bifrankcop, the VGAM family functions for estimating the association parameter by maximum likelihood estimation, for the formula of the cumulative distribution function and other details.

Value

dbifrankcop gives the density, pbifrankcop gives the distribution function, and rbifrankcop generates random deviates (a two-column matrix).

Author(s)

T. W. Yee

References

Genest, C. (1987) Frank's family of bivariate distributions. Biometrika, 74, 549–555.

See Also

bifrankcop.

Examples

## Not run: N <- 100; apar <- exp(2)
xx <- seq(-0.30, 1.30, len = N)
ox <- expand.grid(xx, xx)
zedd <- dbifrankcop(ox[, 1], ox[, 2], apar = apar)
contour(xx, xx, matrix(zedd, N, N))
zedd <- pbifrankcop(ox[, 1], ox[, 2], apar = apar)
contour(xx, xx, matrix(zedd, N, N))

plot(rr <- rbifrankcop(n = 3000, apar = exp(4)))
par(mfrow = c(1, 2))
hist(rr[, 1]); hist(rr[, 2])  # Should be uniform

## End(Not run)

[Package VGAM version 1.0-6 Index]