bilogistic {VGAM}	R Documentation

Bivariate Logistic Distribution Family Function

Description

Estimates the four parameters of the bivariate logistic distribution by maximum likelihood estimation.

Usage

bilogistic(llocation = "identitylink", lscale = "loge",
           iloc1 = NULL, iscale1 = NULL, iloc2 = NULL, iscale2 = NULL,
           imethod = 1, nsimEIM = 250, zero = NULL)

Arguments

llocation
Link function applied to both location parameters l1 and l2. See Links for more choices.

lscale
Parameter link function applied to both (positive) scale parameters s1 and s2. See Links for more choices.

iloc1, iloc2
Initial values for the location parameters. By default, initial values are chosen internally using imethod. Assigning values here will override the argument imethod.

iscale1, iscale2
Initial values for the scale parameters. By default, initial values are chosen internally using imethod. Assigning values here will override the argument imethod.

imethod
An integer with value 1 or 2 which specifies the initialization method. If failure to converge occurs try the other value.

nsimEIM, zero
See CommonVGAMffArguments for details.

Details

The four-parameter bivariate logistic distribution has a density that can be written as

f(y1,y2;l1,s1,l2,s2) = 2 * exp[-(y1-l1)/s1 - (y1-l1)/s1] / [s1 * s2 * ( 1 + exp[-(y1-l1)/s1] + exp[-(y2-l2)/s2] )^3]

where s1>0 and s2>0 are the scale parameters, and l1 and l2 are the location parameters. Each of the two responses are unbounded, i.e., -Inf<y_j<Inf. The mean of Y1 is l1 etc. The fitted values are returned in a 2-column matrix. The cumulative distribution function is

F(y1,y2;l1,s1,l2,s2) = 1 / (1 + exp[-(y1-l1)/s1] + exp[-(y2-l2)/s2])

The marginal distribution of Y1 is

P(Y1 <= y1) = F(y1;l1,s1) = 1 / (1 + exp[-(y1-l1)/s1]).

By default, eta1=l1, eta2=log(s1), eta3=l2, eta4=log(s2) are the linear/additive predictors.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, rrvglm and vgam.

Author(s)

T. W. Yee

References

Gumbel, E. J. (1961) Bivariate logistic distributions. Journal of the American Statistical Association, 56, 335–349.

Castillo, E., Hadi, A. S., Balakrishnan, N. Sarabia, J. S. (2005) Extreme Value and Related Models with Applications in Engineering and Science, Hoboken, NJ, USA: Wiley-Interscience.

See Also

logistic, rbilogis.

Examples

## Not run: 
ymat <- rbilogis(n <- 50, loc1 = 5, loc2 = 7, scale2 = exp(1))
plot(ymat)
bfit <- vglm(ymat ~ 1, family = bilogistic, trace = TRUE)
coef(bfit, matrix = TRUE)
Coef(bfit)
head(fitted(bfit))
vcov(bfit)
head(weights(bfit, type = "work"))
summary(bfit)

## End(Not run)

[Package VGAM version 1.0-6 Index]