binom2.rho {VGAM}	R Documentation

Bivariate Probit Model (Family Function)

Description

Fits a bivariate probit model to two binary responses. The correlation parameter rho is the measure of dependency.

Usage

binom2.rho(lmu = "probit", lrho = "rhobit", imu1 = NULL, imu2 = NULL,
           irho = NULL, imethod = 1, zero = "rho",
           exchangeable = FALSE, grho = seq(-0.95, 0.95, by = 0.05),
           nsimEIM = NULL)
binom2.Rho(rho = 0, imu1 = NULL, imu2 = NULL,
           exchangeable = FALSE, nsimEIM = NULL)

Arguments

lmu
Link function applied to the marginal probabilities. Should be left alone.

lrho
Link function applied to the rho association parameter. See Links for more choices.

imu1, imu2
Optional initial values for the two marginal probabilities. May be a vector.

irho
Optional initial value for rho. If given, this should lie between -1 and 1. See below for more comments.

zero
Specifies which linear/additive predictors are modelled as intercept-only. A NULL means none. Numerically, the rho parameter is easiest modelled as an intercept only, hence the default. See CommonVGAMffArguments for more information.

exchangeable
Logical. If TRUE, the two marginal probabilities are constrained to be equal.

imethod, nsimEIM, grho
See CommonVGAMffArguments for more information. A value of at least 100 for nsimEIM is recommended; the larger the value the better.

rho
Numeric vector. Values are recycled to the needed length, and ought to be in range, which is (-1, 1).

Details

The bivariate probit model was one of the earliest regression models to handle two binary responses jointly. It has a probit link for each of the two marginal probabilities, and models the association between the responses by the rho parameter of a standard bivariate normal distribution (with zero means and unit variances). One can think of the joint probabilities being Phi(eta1,eta2;rho) where Phi is the cumulative distribution function of a standard bivariate normal distribution.

Explicitly, the default model is

probit[P(Y_j=1)] = eta_j, j=1,2

for the marginals, and

rhobit[rho] = eta_3.

The joint probability P(Y_1=1,Y_2=1)=Phi(eta1,eta2;rho), and from these the other three joint probabilities are easily computed. The model is fitted by maximum likelihood estimation since the full likelihood is specified. Fisher scoring is implemented.

The default models eta3 as a single parameter only, i.e., an intercept-only model for rho, but this can be circumvented by setting zero = NULL in order to model rho as a function of all the explanatory variables.

The bivariate probit model should not be confused with a bivariate logit model with a probit link (see binom2.or). The latter uses the odds ratio to quantify the association. Actually, the bivariate logit model is recommended over the bivariate probit model because the odds ratio is a more natural way of measuring the association between two binary responses.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

When fitted, the fitted.values slot of the object contains the four joint probabilities, labelled as (Y1,Y2) = (0,0), (0,1), (1,0), (1,1), respectively.

Note

See binom2.or about the form of input the response should have.

By default, a constant rho is fitted because zero = "rho". Set zero = NULL if you want the rho parameter to be modelled as a function of the explanatory variables. The value rho lies in the interval (-1,1), therefore a rhobit link is default.

Converge problems can occur. If so, assign irho a range of values and monitor convergence (e.g., set trace = TRUE). Else try imethod. Practical experience shows that local solutions can occur, and that irho needs to be quite close to the (global) solution. Also, imu1 and imu2 may be used.

This help file is mainly about binom2.rho(). binom2.Rho() fits a bivariate probit model with known rho. The inputted rho is saved in the misc slot of the fitted object, with rho as the component name.

In some econometrics applications (e.g., Freedman 2010, Freedman and Sekhon 2010) one response is used as an explanatory variable, e.g., a recursive binomial probit model. Such will not work here. Historically, the bivariate probit model was the first VGAM I ever wrote, based on Ashford and Sowden (1970). I don't think they ever thought of it either! Hence the criticisms raised go beyond the use of what was originally intended.

Author(s)

Thomas W. Yee

References

Ashford, J. R. and Sowden, R. R. (1970) Multi-variate probit analysis. Biometrics, 26, 535–546.

Freedman, D. A. (2010) Statistical Models and Causal Inference: a Dialogue with the Social Sciences, Cambridge: Cambridge University Press.

Freedman, D. A. and Sekhon, J. S. (2010) Endogeneity in probit response models. Political Analysis, 18, 138–150.

See Also

rbinom2.rho, rhobit, pbinorm, binom2.or, loglinb2, coalminers, binomialff, rhobit, fisherz.

Examples

coalminers <- transform(coalminers, Age = (age - 42) / 5)
fit <- vglm(cbind(nBnW, nBW, BnW, BW) ~ Age,
            binom2.rho, data = coalminers, trace = TRUE)
summary(fit)
coef(fit, matrix = TRUE)

[Package VGAM version 1.0-6 Index]