Binormcop {VGAM}	R Documentation

Gaussian Copula (Bivariate) Distribution

Description

Density, distribution function, and random generation for the (one parameter) bivariate Gaussian copula distribution.

Usage

dbinormcop(x1, x2, rho = 0, log = FALSE)
pbinormcop(q1, q2, rho = 0)
rbinormcop(n, rho = 0)

Arguments

x1, x2, q1, q2
vector of quantiles. The x1 and x2 should be in the interval (0,1). Ditto for q1 and q2.

n
number of observations. Same as rnorm.

rho
the correlation parameter. Should be in the interval (-1,1).

log
Logical. If TRUE then the logarithm is returned.

Details

See binormalcop, the VGAM family functions for estimating the parameter by maximum likelihood estimation, for the formula of the cumulative distribution function and other details.

Value

dbinormcop gives the density, pbinormcop gives the distribution function, and rbinormcop generates random deviates (a two-column matrix).

Note

Yettodo: allow x1 and/or x2 to have values 1, and to allow any values for x1 and/or x2 to be outside the unit square.

Author(s)

T. W. Yee

See Also

binormalcop, binormal.

Examples

## Not run:  edge <- 0.01  # A small positive value
N <- 101; x <- seq(edge, 1.0 - edge, len = N); Rho <- 0.7
ox <- expand.grid(x, x)
zedd <- dbinormcop(ox[, 1], ox[, 2], rho = Rho, log = TRUE)
contour(x, x, matrix(zedd, N, N), col = "blue", labcex = 1.5)
zedd <- pbinormcop(ox[, 1], ox[, 2], rho = Rho)
contour(x, x, matrix(zedd, N, N), col = "blue", labcex = 1.5)

## End(Not run)

[Package VGAM version 1.0-6 Index]