biplackettcop {VGAM}	R Documentation

Plackett's Bivariate Copula Family Function

Description

Estimate the association parameter of Plackett's bivariate distribution (copula) by maximum likelihood estimation.

Usage

biplackettcop(link = "loge", ioratio = NULL, imethod = 1, nsimEIM = 200)

Arguments

link
Link function applied to the (positive) odds ratio psi. See Links for more choices and information.

ioratio
Numeric. Optional initial value for psi. If a convergence failure occurs try assigning a value or a different value.

imethod, nsimEIM
See CommonVGAMffArguments.

Details

The defining equation is

psi = H*(1-y1-y2+H) / ((y1-H)*(y2-H))

where P(Y1 <= y1, Y2 <= y2)= H(y1,y2) is the cumulative distribution function. The density function is h(y1,y2) =

psi*[1 + (psi-1)*(y1 + y2 - 2*y1*y2) ] / ( [1 + (psi-1)*(y1 + y2)]^2 - 4*psi*(psi-1)*y1*y2)^(3/2)

for psi > 0. Some writers call psi the cross product ratio but it is called the odds ratio here. The support of the function is the unit square. The marginal distributions here are the standard uniform although it is commonly generalized to other distributions.

If psi=1 then h(y1,y2) = y1*y2, i.e., independence. As the odds ratio tends to infinity one has y1=y2. As the odds ratio tends to 0 one has y2=1-y1.

Fisher scoring is implemented using rbiplackcop. Convergence is often quite slow.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Note

The response must be a two-column matrix. Currently, the fitted value is a 2-column matrix with 0.5 values because the marginal distributions correspond to a standard uniform distribution.

Author(s)

T. W. Yee

References

Plackett, R. L. (1965) A class of bivariate distributions. Journal of the American Statistical Association, 60, 516–522.

See Also

rbiplackcop, bifrankcop.

Examples

## Not run: 
ymat <- rbiplackcop(n = 2000, oratio = exp(2))
plot(ymat, col = "blue")
fit <- vglm(ymat ~ 1, fam = biplackettcop, trace = TRUE)
coef(fit, matrix = TRUE)
Coef(fit)
vcov(fit)
head(fitted(fit))
summary(fit)

## End(Not run)

[Package VGAM version 1.0-6 Index]