biplot-methods {VGAM}	R Documentation

Biplot of Constrained Regression Models

Description

biplot is a generic function applied to RR-VGLMs and QRR-VGLMs etc. These apply to rank-1 and rank-2 models of these only. For RR-VGLMs these plot the second latent variable scores against the first latent variable scores.

Methods

x

The object from which the latent variables are extracted and/or plotted.

Note

See lvplot which is very much related to biplots.

[Package VGAM version 1.0-6 Index]