Bistudentt {VGAM}	R Documentation

Bivariate Student-t Distribution Cumulative Distribution Function

Description

Density for the bivariate Student-t distribution distribution.

Usage

dbistudentt(x1, x2, df, rho = 0, log = FALSE)

Arguments

x1, x2
vector of quantiles.

df, rho
vector of degrees of freedom and correlation parameter. For df, a value Inf is currently not working.

log
Logical. If log = TRUE then the logarithm of the density is returned.

Details

One can think of this function as an extension of dt to two dimensions. See bistudentt for more information.

Value

dbistudentt gives the density.

References

Schepsmeier, U. and Stober, J. (2013) Derivatives and Fisher information of bivariate copulas. Statistical Papers.

See Also

bistudentt, dt.

Examples

## Not run:  N <- 101; x <- seq(-4, 4, len = N); Rho <- 0.7; mydf <- 10
ox <- expand.grid(x, x)
zedd <- dbistudentt(ox[, 1], ox[, 2], df = mydf, rho = Rho, log = TRUE)
contour(x, x, matrix(zedd, N, N), col = "blue", labcex = 1.5)

## End(Not run)

[Package VGAM version 1.0-6 Index]