bmi.nz {VGAM}	R Documentation

Body Mass Index of New Zealand Adults Data

Description

The body mass indexes and ages from an approximate random sample of 700 New Zealand adults.

Usage

data(bmi.nz)

Format

A data frame with 700 observations on the following 2 variables.

age

a numeric vector; their age (years).

BMI

a numeric vector; their body mass indexes, which is their weight divided by the square of their height (kg / m^2).

Details

They are a random sample from the Fletcher Challenge/Auckland Heart and Health survey conducted in the early 1990s.

There are some outliers in the data set.

A variable gender would be useful, and may be added later.

Source

Clinical Trials Research Unit, University of Auckland, New Zealand, http://www.ctru.auckland.ac.nz.

References

MacMahon, S., Norton, R., Jackson, R., Mackie, M. J., Cheng, A., Vander Hoorn, S., Milne, A., McCulloch, A. (1995) Fletcher Challenge-University of Auckland Heart & Health Study: design and baseline findings. New Zealand Medical Journal, 108, 499–502.

Examples

## Not run:  with(bmi.nz, plot(age, BMI, col = "blue"))
fit <- vgam(BMI ~ s(age, df = c(2, 4, 2)), lms.yjn, data = bmi.nz, trace = TRUE)
qtplot(fit, pcol = "blue", tcol = "brown", lcol = "brown") 
## End(Not run)

[Package VGAM version 1.0-6 Index]