borel.tanner {VGAM}	R Documentation

Borel-Tanner Distribution Family Function

Description

Estimates the parameter of a Borel-Tanner distribution by maximum likelihood estimation.

Usage

borel.tanner(Qsize = 1, link = "logit", imethod = 1)

Arguments

Qsize
A positive integer. It is called Q below and is the initial queue size. The default value Q = 1 corresponds to the Borel distribution.

link
Link function for the parameter; see Links for more choices and for general information.

imethod
See CommonVGAMffArguments. Valid values are 1, 2, 3 or 4.

Details

The Borel-Tanner distribution (Tanner, 1953) describes the distribution of the total number of customers served before a queue vanishes given a single queue with random arrival times of customers (at a constant rate r per unit time, and each customer taking a constant time b to be served). Initially the queue has Q people and the first one starts to be served. The two parameters appear in the density only in the form of the product rb, therefore we use a=rb, say, to denote the single parameter to be estimated. The density function is

f(y;a) = (Q / (y-Q)!) * y^(y-Q-1) * a^(y-Q) * exp(-ay)

where y=Q,Q+1,Q+2,.... The case Q=1 corresponds to the Borel distribution (Borel, 1942). For the Q=1 case it is necessary for 0 < a < 1 for the distribution to be proper. The Borel distribution is a basic Lagrangian distribution of the first kind. The Borel-Tanner distribution is an Q-fold convolution of the Borel distribution.

The mean is Q/(1-a) (returned as the fitted values) and the variance is Q*a/(1-a)^3. The distribution has a very long tail unless a is small. Fisher scoring is implemented.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Author(s)

T. W. Yee

References

Tanner, J. C. (1953) A problem of interference between two queues. Biometrika, 40, 58–69.

Borel, E. (1942) Sur l'emploi du theoreme de Bernoulli pour faciliter le calcul d'une infinite de coefficients. Application au probleme de l'attente a un guichet. Comptes Rendus, Academie des Sciences, Paris, Series A, 214, 452–456.

Page 328 of Johnson N. L., Kemp, A. W. and Kotz S. (2005) Univariate Discrete Distributions, 3rd edition, Hoboken, New Jersey: Wiley.

Consul, P. C. and Famoye, F. (2006) Lagrangian Probability Distributions, Boston, MA, USA: Birkhauser.

See Also

rbort, poissonff, felix.

Examples

bdata <- data.frame(y = rbort(n <- 200))
fit <- vglm(y ~ 1, borel.tanner, data = bdata, trace = TRUE, crit = "c")
coef(fit, matrix = TRUE)
Coef(fit)
summary(fit)

[Package VGAM version 1.0-6 Index]