Bort {VGAM}	R Documentation

The Borel-Tanner Distribution

Description

Density and random generation for the Borel-Tanner distribution.

Usage

dbort(x, Qsize = 1, a = 0.5, log = FALSE)
rbort(n, Qsize = 1, a = 0.5)

Arguments

x
vector of quantiles.

n
number of observations. Must be a positive integer of length 1.

Qsize, a
See borel.tanner.

log
Logical. If log = TRUE then the logarithm of the density is returned.

Details

See borel.tanner, the VGAM family function for estimating the parameter, for the formula of the probability density function and other details.

Value

dbort gives the density, rbort generates random deviates.

Warning

Looping is used for rbort, therefore values of a close to 1 will result in long (or infinite!) computational times. The default value of a is subjective.

Author(s)

T. W. Yee

See Also

borel.tanner.

Examples

## Not run:  qsize <- 1; a <- 0.5; x <- qsize:(qsize+10)
plot(x, dbort(x, qsize, a), type = "h", las = 1, col = "blue",
     ylab = paste("fbort(qsize=", qsize, ", a=", a, ")"), log = "y",
     main = "Borel-Tanner density function") 
## End(Not run)

[Package VGAM version 1.0-6 Index]