calibrate {VGAM}	R Documentation

Model Calibrations

Description

calibrate is a generic function used to produce calibrations from various model fitting functions. The function invokes particular ‘methods’ which depend on the ‘class’ of the first argument.

Usage

calibrate(object, ...)

Arguments

object
An object for which a calibration is desired.

...
Additional arguments affecting the calibration produced. Usually the most important argument in ... is newdata which, for calibrate, contains new response data, Y, say.

Details

Given a regression model with explanatory variables X and response Y, calibration involves estimating X from Y using the regression model. It can be loosely thought of as the opposite of predict (which takes an X and returns a Y of some sort.)

Value

In general, given a new response Y, some function of the explanatory variables X are returned. For example, for constrained ordination models such as CQO and CAO models, it is usually not possible to return X, so the latent variables are returned instead (they are linear combinations of the X). See the specific calibrate methods functions to see what they return.

Note

This function was not called predictx because of the inability of constrained ordination models to return X; they can only return the latent variable values (also known as site scores) instead.

Author(s)

T. W. Yee

See Also

predict, calibrate.rrvglm, calibrate.qrrvglm.

Examples

## Not run: 
hspider[, 1:6] <- scale(hspider[, 1:6])  # Stdzed environmental vars
set.seed(123)
pcao1 <- cao(cbind(Pardlugu, Pardmont, Pardnigr, Pardpull, Zoraspin) ~
         WaterCon + BareSand + FallTwig + CoveMoss + CoveHerb + ReflLux,
         family = poissonff, data = hspider, Rank = 1, Bestof = 3,
         df1.nl = c(Zoraspin = 2, 1.9), Crow1positive = TRUE)

siteNos <- 1:2  # Calibrate these sites
cpcao1 <- calibrate(pcao1, trace = TRUE,
                    newdata = data.frame(depvar(pcao1)[siteNos, ],
                                         model.matrix(pcao1)[siteNos, ]))

# Graphically compare the actual site scores with their calibrated values
persp(pcao1, main = "Site scores: solid=actual, dashed=calibrated",
      label = TRUE, col = "blue", las = 1)
abline(v = latvar(pcao1)[siteNos], col = seq(siteNos))  # Actual scores
abline(v = cpcao1, lty = 2, col = seq(siteNos))  # Calibrated values

## End(Not run)

[Package VGAM version 1.0-6 Index]