calibrate.rrvglm.control {VGAM}	R Documentation

Control Function for CLO (RR-VGLM) Calibration

Description

Algorithmic constants and parameters for running calibrate.rrvglm are set using this function.

Usage

calibrate.rrvglm.control(object, trace = FALSE, method.optim = "BFGS",
    gridSize = ifelse(Rank == 1, 17, 9), ...)

Arguments

object
The fitted rrvglm model. The user should ignore this argument.

trace, method.optim
Same as calibrate.qrrvglm.control.

gridSize
Same as calibrate.qrrvglm.control.

...
Avoids an error message for extraneous arguments.

Details

Most CLO users will only need to make use of trace and gridSize. These arguments should be used inside their call to calibrate.rrvglm, not this function directly.

Value

Similar to calibrate.qrrvglm.control.

See Also

calibrate.rrvglm, Coef.rrvglm.

[Package VGAM version 1.0-6 Index]