cfibrosis {VGAM}	R Documentation

Cystic Fibrosis Data

Description

This data frame concerns families data and cystic fibrosis.

Usage

data(cfibrosis)

Format

A data frame with 24 rows on the following 4 variables.

siblings, affected, ascertained, families

Over ascertained families, the kth ascertained family has s_k siblings of whom r_k are affected and a_k are ascertained.

Details

The data set allows a classical segregation analysis to be peformed. In particular, to test Mendelian segregation ratios in nuclear family data. The likelihood has similarities with seq2binomial.

Source

The data is originally from Crow (1965) and appears as Table 2.3 of Lange (2002).

Crow, J. F. (1965) Problems of ascertainment in the analysis of family data. Epidemiology and Genetics of Chronic Disease. Public Health Service Publication 1163, Neel J. V., Shaw M. W., Schull W. J., editors, Department of Health, Education, and Welfare, Washington, DC, USA.

Lange, K. (2002) Mathematical and Statistical Methods for Genetic Analysis. Second Edition. Springer-Verlag: New York, USA.

Examples

cfibrosis
summary(cfibrosis)

[Package VGAM version 1.0-6 Index]