chest.nz {VGAM}	R Documentation

Chest Pain in NZ Adults Data

Description

Presence/absence of chest pain in 10186 New Zealand adults.

Usage

data(chest.nz)

Format

A data frame with 73 rows and the following 5 variables.

age

a numeric vector; age (years).

nolnor

a numeric vector of counts; no pain on LHS or RHS.

nolr

a numeric vector of counts; no pain on LHS but pain on RHS.

lnor

a numeric vector of counts; no pain on RHS but pain on LHS.

lr

a numeric vector of counts; pain on LHS and RHS of chest.

Details

Each adult was asked their age and whether they experienced any pain or discomfort in their chest over the last six months. If yes, they indicated whether it was on their LHS and/or RHS of their chest.

Source

MacMahon, S., Norton, R., Jackson, R., Mackie, M. J., Cheng, A., Vander Hoorn, S., Milne, A., McCulloch, A. (1995) Fletcher Challenge-University of Auckland Heart & Health Study: design and baseline findings. New Zealand Medical Journal, 108, 499–502.

Examples

## Not run: 
fit <- vgam(cbind(nolnor, nolr, lnor, lr) ~ s(age, c(4, 3)),
            binom2.or(exchan = TRUE, zero = NULL), data = chest.nz)
coef(fit, matrix = TRUE)

## End(Not run)
## Not run:  plot(fit, which.cf = 2, se = TRUE) 

[Package VGAM version 1.0-6 Index]