clo {VGAM}	R Documentation

Redirects the User to rrvglm()

Description

Redirects the user to the function rrvglm.

Usage

clo(...)

Arguments

...
Ignored.

Details

CLO stands for constrained linear ordination, and is fitted with a statistical class of models called reduced-rank vector generalized linear models (RR-VGLMs). It allows for generalized reduced-rank regression in that response types such as Poisson counts and presence/absence data can be handled.

Currently in the VGAM package, rrvglm is used to fit RR-VGLMs. However, the Author's opinion is that linear responses to a latent variable (composite environmental gradient) is not as common as unimodal responses, therefore cqo is often more appropriate.

The new CLO/CQO/CAO nomenclature described in Yee (2006).

Value

Nothing is returned; an error message is issued.

Author(s)

Thomas W. Yee

References

Yee, T. W. (2006) Constrained additive ordination. Ecology, 87, 203–213.

Yee, T. W. and Hastie, T. J. (2003) Reduced-rank vector generalized linear models. Statistical Modelling, 3, 15–41.

See Also

rrvglm, cqo.

Examples

## Not run: 
clo()

## End(Not run)

[Package VGAM version 1.0-6 Index]