coalminers {VGAM}	R Documentation

Breathlessness and Wheeze Amongst Coalminers Data

Description

Coalminers who are smokers without radiological pneumoconiosis, classified by age, breathlessness and wheeze.

Usage

data(coalminers)

Format

A data frame with 9 age groups with the following 5 columns.

BW

Counts with breathlessness and wheeze.

BnW

Counts with breathlessness but no wheeze.

nBW

Counts with no breathlessness but wheeze.

nBnW

Counts with neither breathlessness or wheeze.

age

Age of the coal miners (actually, the midpoints of the 5-year category ranges).

Details

The data were published in Ashford and Sowden (1970). A more recent analysis is McCullagh and Nelder (1989, Section 6.6).

Source

Ashford, J. R. and Sowden, R. R. (1970) Multi-variate probit analysis. Biometrics, 26, 535–546.

References

McCullagh, P. and Nelder, J. A. (1989) Generalized Linear Models. 2nd ed. London: Chapman & Hall.

Examples

str(coalminers)

[Package VGAM version 1.0-6 Index]