concoef {VGAM}	R Documentation

Extract Model Constrained/Canonical Coefficients

Description

concoef is a generic function which extracts the constrained (canonical) coefficients from objects returned by certain modelling functions.

Usage

concoef(object, ...)

Arguments

object
An object for which the extraction of canonical coefficients is meaningful.

...
Other arguments fed into the specific methods function of the model.

Details

For constrained quadratic and ordination models, canonical coefficients are the elements of the C matrix used to form the latent variables. They are highly interpretable in ecology, and are looked at as weights or loadings.

They are also applicable for reduced-rank VGLMs.

Value

The value returned depends specifically on the methods function invoked.

Warning

concoef replaces ccoef; the latter is deprecated.

For QO models, there is a direct inverse relationship between the scaling of the latent variables (site scores) and the tolerances. One normalization is for the latent variables to have unit variance. Another normalization is for all the species' tolerances to be unit (provided eq.tolerances is TRUE). These two normalizations cannot simultaneously hold in general. For rank R models with R>1 it becomes more complicated because the latent variables are also uncorrelated. An important argument when fitting quadratic ordination models is whether eq.tolerances is TRUE or FALSE. See Yee (2004) for details.

Author(s)

Thomas W. Yee

References

Yee, T. W. and Hastie, T. J. (2003) Reduced-rank vector generalized linear models. Statistical Modelling, 3, 15–41.

Yee, T. W. (2004) A new technique for maximum-likelihood canonical Gaussian ordination. Ecological Monographs, 74, 685–701.

Yee, T. W. (2006) Constrained additive ordination. Ecology, 87, 203–213.

See Also

concoef-method, concoef.qrrvglm, concoef.cao, coef.

Examples

## Not run:  set.seed(111)  # This leads to the global solution
hspider[,1:6] <- scale(hspider[,1:6])  # Standardized environmental vars
p1 <- cqo(cbind(Alopacce, Alopcune, Alopfabr, Arctlute, Arctperi,
                Auloalbi, Pardlugu, Pardmont, Pardnigr, Pardpull,
                Trocterr, Zoraspin) ~
          WaterCon + BareSand + FallTwig + CoveMoss + CoveHerb + ReflLux,
          family = poissonff, data = hspider, Crow1positive = FALSE)
concoef(p1)

## End(Not run)

[Package VGAM version 1.0-6 Index]