concoef-methods {VGAM}	R Documentation

Constrained (Canonical) Coefficients

Description

concoef is a generic function used to return the constrained (canonical) coefficients of a constrained ordination model. The function invokes particular methods which depend on the class of the first argument.

Methods

object

The object from which the constrained coefficients are extracted.

[Package VGAM version 1.0-6 Index]