corbet {VGAM}	R Documentation

Corbet's Butterfly Data

Description

About 3300 individual butterflies were caught in Malaya by naturalist Corbet trapping butterflies. They were classified to about 500 species.

Usage

data(corbet)

Format

A data frame with 24 observations on the following 2 variables.

species

Number of species.

ofreq

Observed frequency of individual butterflies of that species.

Details

In the early 1940s Corbet spent two years trapping butterflies in Malaya. Of interest was the total number of species. Some species were so rare (e.g., 118 species had only one specimen) that it was thought likely that there were many unknown species.

References

Fisher, R. A., Corbet, A. S. and Williams, C. B. (1943) The Relation Between the Number of Species and the Number of Individuals in a Random Sample of an Animal Population. Journal of Animal Ecology, 12, 42–58.

Examples

summary(corbet)

[Package VGAM version 1.0-6 Index]