deermice {VGAM}	R Documentation

Captures of Peromyscus maniculatus (Also Known as Deer Mice).

Description

Captures of Peromyscus maniculatus collected at East Stuart Gulch, Colorado, USA.

Usage

  data(deermice)

Format

The format is a data frame.

Details

Peromyscus maniculatus is a rodent native to North America. The deer mouse is small in size, only about 8 to 10 cm long, not counting the length of the tail.

Originally, the columns of this data frame represent the sex (m or f), the ages (y: young, sa: semi-adult, a: adult), the weights in grams, and the capture histories of 38 individuals over 6 trapping occasions (1: captured, 0: not captured).

The data set was collected by V. Reid and distributed with the CAPTURE program of Otis et al. (1978).

deermice has 38 deermice whereas Perom had 36 deermice (Perom has been withdrawn.) In deermice the two semi-adults have been classified as adults. The sex variable has 1 for female, and 0 for male.

References

Huggins, R. M. (1991) Some practical aspects of a conditional likelihood approach to capture experiments. Biometrics, 47, 725–732.

Otis, D. L. et al. (1978) Statistical inference from capture data on closed animal populations, Wildlife Monographs, 62, 3–135.

See Also

posbernoulli.b, posbernoulli.t, fill1.

Examples

head(deermice)
## Not run: 
fit1 <- vglm(cbind(y1, y2, y3, y4, y5, y6) ~ sex + age,
             posbernoulli.t(parallel.t = TRUE), data = deermice, trace = TRUE)
coef(fit1)
coef(fit1, matrix = TRUE)

## End(Not run)

[Package VGAM version 1.0-6 Index]