diffzeta {VGAM}	R Documentation

Differenced Zeta Distribution Family Function

Description

Estimates the parameter of the differenced zeta distribution.

Usage

diffzeta(start = 1, lshape = "loge", ishape = NULL)

Arguments

lshape, ishape
Same as zetaff.

start
Smallest value of the support of the distribution. Must be a positive integer.

Details

The PMF is

P(Y=y) = (a/y)^(s) - / (a/(1+y))^(s), s>0, y=a,a+1,...,

where s is the positive shape parameter, and a is start. According to Moreno-Sanchez et al. (2016), this model fits quite well to about 40 percent of all the English books in the Project Gutenberg data base (about 30,000 texts). Multiple responses are handled.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Author(s)

T. W. Yee

References

Moreno-Sanchez, I. and Font-Clos, F. and Corral, A. Large-Scale Analysis of Zipf's Law in English Texts, 2016. PLoS ONE, 11(1), 1–19.

See Also

Diffzeta, zetaff, zeta, zipf, zipf.

Examples

odata <- data.frame(x2 = runif(nn <- 1000))  # Artificial data
odata <- transform(odata, shape = loge(-0.25 + x2, inverse = TRUE))
odata <- transform(odata, y1 = rdiffzeta(nn, shape))
with(odata, table(y1))
ofit <- vglm(y1 ~ x2, diffzeta, data = odata, trace = TRUE, crit = "coef")
coef(ofit, matrix = TRUE)

[Package VGAM version 1.0-6 Index]