ducklings {VGAM}	R Documentation

Relative Frequencies of Serum Proteins in White Pekin Ducklings

Description

Relative frequencies of serum proteins in white Pekin ducklings as determined by electrophoresis.

Usage

data(ducklings)

Format

The format is: chr "ducklings"

Details

Columns p1, p2, p3 stand for pre-albumin, albumin, globulins respectively. These were collected from 3-week old white Pekin ducklings. Let Y1 be proportional to the total milligrams of pre-albumin in the blood serum of a duckling. Similarly, let Y2 and Y3 be directly proportional to the same factor as Y1 to the total milligrams respectively of albumin and globulins in its blood serum. The proportion of pre-albumin is given by Y1/(Y1 + Y2 + Y3), and similarly for the others.

Source

Mosimann, J. E. (1962) On the compound multinomial distribution, the multivariate beta-distribution, and correlations among proportions, Biometrika, 49, 65–82.

See Also

dirichlet.

Examples

print(ducklings)

[Package VGAM version 1.0-6 Index]