expexpff1 {VGAM}	R Documentation

Exponentiated Exponential Distribution

Description

Estimates the two parameters of the exponentiated exponential distribution by maximizing a profile (concentrated) likelihood.

Usage

expexpff1(lrate = "loge", irate = NULL, ishape = 1)

Arguments

lrate
Parameter link function for the (positive) rate parameter. See Links for more choices.

irate
Initial value for the rate parameter. By default, an initial value is chosen internally using ishape.

ishape
Initial value for the shape parameter. If convergence fails try setting a different value for this argument.

Details

See expexpff for details about the exponentiated exponential distribution. This family function uses a different algorithm for fitting the model. Given rate, the MLE of shape can easily be solved in terms of rate. This family function maximizes a profile (concentrated) likelihood with respect to rate. Newton-Raphson is used, which compares with Fisher scoring with expexpff.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Warning

The standard errors produced by a summary of the model may be wrong.

Note

This family function works only for intercept-only models, i.e., y ~ 1 where y is the response.

The estimate of shape is attached to the misc slot of the object, which is a list and contains the component shape.

As Newton-Raphson is used, the working weights are sometimes negative, and some adjustment is made to these to make them positive.

Like expexpff, good initial values are needed. Convergence may be slow.

Author(s)

T. W. Yee

References

Gupta, R. D. and Kundu, D. (2001) Exponentiated exponential family: an alternative to gamma and Weibull distributions, Biometrical Journal, 43, 117–130.

See Also

expexpff, CommonVGAMffArguments.

Examples

# Ball bearings data (number of million revolutions before failure)
edata <- data.frame(bbearings = c(17.88, 28.92, 33.00, 41.52, 42.12, 45.60,
48.80, 51.84, 51.96, 54.12, 55.56, 67.80, 68.64, 68.64,
68.88, 84.12, 93.12, 98.64, 105.12, 105.84, 127.92,
128.04, 173.40))
fit <- vglm(bbearings ~ 1, expexpff1(ishape = 4), trace = TRUE,
            maxit = 250, checkwz = FALSE, data = edata)
coef(fit, matrix = TRUE)
Coef(fit)  # Authors get c(0.0314, 5.2589) with log-lik -112.9763
logLik(fit)
fit@misc$shape  # Estimate of shape

# Failure times of the airconditioning system of an airplane
eedata <- data.frame(acplane = c(23, 261, 87, 7, 120, 14, 62, 47,
225, 71, 246, 21, 42, 20, 5, 12, 120, 11, 3, 14,
71, 11, 14, 11, 16, 90, 1, 16, 52, 95))
fit <- vglm(acplane ~ 1, expexpff1(ishape = 0.8), trace = TRUE,
            maxit = 50, checkwz = FALSE, data = eedata)
coef(fit, matrix = TRUE)
Coef(fit)  # Authors get c(0.0145, 0.8130) with log-lik -152.264
logLik(fit)
fit@misc$shape  # Estimate of shape

[Package VGAM version 1.0-6 Index]