expgeometric {VGAM}	R Documentation

Exponential Geometric Distribution Family Function

Description

Estimates the two parameters of the exponential geometric distribution by maximum likelihood estimation.

Usage

expgeometric(lscale = "loge", lshape = "logit",
             iscale = NULL,   ishape = NULL,
             tol12 = 1e-05, zero = 1, nsimEIM = 400)

Arguments

lscale, lshape
Link function for the two parameters. See Links for more choices.

iscale, ishape
Numeric. Optional initial values for the scale and shape parameters.

tol12
Numeric. Tolerance for testing whether a parameter has value 1 or 2.

zero, nsimEIM
See CommonVGAMffArguments.

Details

The exponential geometric distribution has density function

(1/c) * (1 - s) * e^(-y/c) * (1 - s * e^(-y/c))^(-2)

where y > 0, c > 0 and 0 < s < 1. The mean, (c * (s - 1)/ s) * log(1 - s) is returned as the fitted values. Note the median is c * log(2 - s). Simulated Fisher scoring is implemented.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Note

We define scale as the reciprocal of the scale parameter used by Adamidis and Loukas (1998).

Author(s)

J. G. Lauder and T. W. Yee

References

Adamidis, K., Loukas, S. (1998). A lifetime distribution with decreasing failure rate. Statistics and Probability Letters, 39, 35–42.

See Also

dexpgeom, exponential, geometric.

Examples

## Not run: 
Scale <- exp(2); shape = logit(-1, inverse = TRUE);
edata <- data.frame(y = rexpgeom(n = 2000, scale = Scale, shape = shape))
fit <- vglm(y ~ 1, expgeometric, edata, trace = TRUE)
c(with(edata, mean(y)), head(fitted(fit), 1))
coef(fit, matrix = TRUE)
Coef(fit)
summary(fit)

## End(Not run)

[Package VGAM version 1.0-6 Index]