explogff {VGAM}	R Documentation

Exponential Logarithmic Distribution Family Function

Description

Estimates the two parameters of the exponential logarithmic distribution by maximum likelihood estimation.

Usage

explogff(lscale = "loge", lshape = "logit",
         iscale = NULL,   ishape = NULL,
         tol12 = 1e-05, zero = 1, nsimEIM = 400)

Arguments

lscale, lshape
See CommonVGAMffArguments for information.

tol12
Numeric. Tolerance for testing whether a parameter has value 1 or 2.

iscale, ishape, zero, nsimEIM
See CommonVGAMffArguments.

Details

The exponential logarithmic distribution has density function

(1/(-log(p))) * (((1/c) * (1 - s) * e^(-y/c)) / (1 - (1 - s) * e^(-y/c)))

where y > 0, scale parameter c > 0, and shape parameter 0 < s < 1. The mean, ((-polylog(2, 1 - p) * c) / log(s) is not returned as the fitted values. Note the median is c * log(1 + sqrt(s)) and it is currently returned as the fitted values. Simulated Fisher scoring is implemented.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Note

We define scale as the reciprocal of the rate parameter used by Tahmasabi and Sadegh (2008).

Yet to do: find a polylog() function.

Author(s)

J. G. Lauder and T. W .Yee

References

Tahmasabi, R., Sadegh, R. (2008). A two-parameter lifetime distribution with decreasing failure rate. Computational Statistics and Data Analysis, 52, 3889–3901.

See Also

dexplog, exponential,

Examples

## Not run:  Scale <- exp(2); shape <- logit(-1, inverse = TRUE)
edata <- data.frame(y = rexplog(n = 2000, scale = Scale, shape = shape))
fit <- vglm(y ~ 1, explogff, data = edata, trace = TRUE)
c(with(edata, median(y)), head(fitted(fit), 1))
coef(fit, matrix = TRUE)
Coef(fit)
summary(fit)

## End(Not run)

[Package VGAM version 1.0-6 Index]