exppoisson {VGAM}	R Documentation

Exponential Poisson Distribution Family Function

Description

Estimates the two parameters of the exponential Poisson distribution by maximum likelihood estimation.

Usage

exppoisson(lrate = "loge", lshape = "loge",
           irate = 2, ishape = 1.1,  zero = NULL)

Arguments

lshape, lrate
Link function for the two positive parameters. See Links for more choices.

ishape, irate
Numeric. Initial values for the shape and rate parameters. Currently this function is not intelligent enough to obtain better initial values.

zero
See CommonVGAMffArguments.

Details

The exponential Poisson distribution has density function

f(y; a = shape, b = rate) = (a*b/(1 - e^(-a))) * e^{-a - b*y + a * e^(-b*y)}

where y > 0, and the parameters shape, a, and rate, b, are positive. The distribution implies a population facing discrete hazard rates which are multiples of a base hazard. This VGAM family function requires the hypergeo package (to use their genhypergeo function). The median is returned as the fitted value.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Warning

This VGAM family function does not work properly!

Author(s)

J. G. Lauder, jamesglauder@gmail.com

References

Kus, C., (2007). A new lifetime distribution. Computational Statistics and Data Analysis, 51, 4497–4509.

See Also

dexppois, exponential, poisson.

Examples

## Not run: 
shape <- exp(1); rate <- exp(2)
rdata <- data.frame(y = rexppois(n = 1000, rate = rate, shape = shape))
library("hypergeo")  # Required!
fit <- vglm(y ~ 1, exppoisson, data = rdata, trace = FALSE, maxit = 1200)
c(with(rdata, median(y)), head(fitted(fit), 1))
coef(fit, matrix = TRUE)
Coef(fit)
summary(fit)

## End(Not run)

[Package VGAM version 1.0-6 Index]