familyname {VGAM}	R Documentation

Family Function Name

Description

Extractor function for the name of the family function of an object in the VGAM package.

Usage

familyname(object, ...)
familyname.vlm(object, all = FALSE, ...)

Arguments

object
Some VGAM object, for example, having class vglmff-class.

all
If all = TRUE then all of the vfamily slot is returned; this contains subclasses the object might have. The default is the return the first value only.

...
Other possible arguments for the future.

Details

Currently VGAM implements over 150 family functions. This function returns the name of the function assigned to the family argument, for modelling functions such as vglm and vgam. Sometimes a slightly different answer is returned, e.g., propodds really calls cumulative with some arguments set, hence the output returned by this function is "cumulative" (note that one day this might change, however).

Value

A character string or vector.

Note

Arguments used in the invocation are not included. Possibly this is something to be done in the future.

See Also

vglmff-class, vglm-class.

Examples

pneumo <- transform(pneumo, let = log(exposure.time))
fit1 <- vglm(cbind(normal, mild, severe) ~ let,
              cumulative(parallel = TRUE, reverse = TRUE), data = pneumo)
familyname(fit1)
familyname(fit1, all = TRUE)
familyname(propodds())  # "cumulative"

[Package VGAM version 1.0-6 Index]