felix {VGAM}	R Documentation

Felix Distribution Family Function

Description

Estimates the parameter of a Felix distribution by maximum likelihood estimation.

Usage

felix(lrate = extlogit(min = 0, max = 0.5), imethod = 1)

Arguments

lrate
Link function for the parameter, called a below; see Links for more choices and for general information.

imethod
See CommonVGAMffArguments. Valid values are 1, 2, 3 or 4.

Details

The Felix distribution is an important basic Lagrangian distribution. The density function is

f(y;a) = (1 / ((y-1)/2)!) * y^((y-3)/2) * a^((y-1)/2) * exp(-ay)

where y=1,3,5,… and 0 < a < 0.5. The mean is 1/(1-2a) (returned as the fitted values). Fisher scoring is implemented.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Author(s)

T. W. Yee

References

Consul, P. C. and Famoye, F. (2006) Lagrangian Probability Distributions, Boston, USA: Birkhauser.

See Also

dfelix, borel.tanner.

Examples

fdata <- data.frame(y = 2 * rpois(n = 200, 1) + 1)  # Not real data!
fit <- vglm(y ~ 1, felix, data = fdata, trace = TRUE, crit = "coef")
coef(fit, matrix = TRUE)
Coef(fit)
summary(fit)

[Package VGAM version 1.0-6 Index]