Felix {VGAM}	R Documentation

The Felix Distribution

Description

Density for the Felix distribution.

Usage

dfelix(x, rate = 0.25, log = FALSE)

Arguments

x
vector of quantiles.

rate
See felix.

log
Logical. If log = TRUE then the logarithm of the density is returned.

Details

See felix, the VGAM family function for estimating the parameter, for the formula of the probability density function and other details.

Value

dfelix gives the density.

Warning

The default value of rate is subjective.

Author(s)

T. W. Yee

See Also

felix.

Examples

## Not run: 
rate <- 0.25; x <- 1:15
plot(x, dfelix(x, rate), type = "h", las = 1, col = "blue",
     ylab = paste("dfelix(rate=", rate, ")"),
     main = "Felix density function")

## End(Not run)

[Package VGAM version 1.0-6 Index]