finney44 {VGAM}	R Documentation

Toxicity trial for insects

Description

A data frame of a toxicity trial.

Usage

data(finney44)

Format

A data frame with 6 observations on the following 3 variables.

pconc

a numeric vector, percent concentration of pyrethrins.

hatched

number of eggs that hatched.

unhatched

number of eggs that did not hatch.

Details

Finney (1944) describes a toxicity trial of five different concentrations of pyrethrins (percent) plus a control that were administered to eggs of Ephestia kuhniella. The natural mortality rate is large, and a common adjustment is to use Abbott's formula.

References

Finney, D. J., 1944. The application of the probit method to toxicity test data adjusted for mortality in the controls. Annals of Applied Biology, 31, 68–74.

Abbott, W. S. (1925). A method of computing the effectiveness of an insecticide. Journal of Economic Entomology, 18, 265–7.

Examples

data(finney44)
transform(finney44, mortality = unhatched / (hatched + unhatched))

[Package VGAM version 1.0-6 Index]