fisk {VGAM}	R Documentation

Fisk Distribution family function

Description

Maximum likelihood estimation of the 2-parameter Fisk distribution.

Usage

fisk(lscale = "loge", lshape1.a = "loge", iscale = NULL,
    ishape1.a = NULL, imethod = 1, lss = TRUE, gscale = exp(-5:5),
    gshape1.a = seq(0.75, 4, by = 0.25), probs.y = c(0.25, 0.5, 0.75),
    zero = "shape")

Arguments

lss
See CommonVGAMffArguments for important information.

lshape1.a, lscale
Parameter link functions applied to the (positive) parameters a and scale. See Links for more choices.

iscale, ishape1.a, imethod, zero
See CommonVGAMffArguments for information. For imethod = 2 a good initial value for iscale is needed to obtain a good estimate for the other parameter.

gscale, gshape1.a
See CommonVGAMffArguments for information.

probs.y
See CommonVGAMffArguments for information.

Details

The 2-parameter Fisk (aka log-logistic) distribution is the 4-parameter generalized beta II distribution with shape parameter q=p=1. It is also the 3-parameter Singh-Maddala distribution with shape parameter q=1, as well as the Dagum distribution with p=1. More details can be found in Kleiber and Kotz (2003).

The Fisk distribution has density

f(y) = a y^(a-1) / [b^a (1 + (y/b)^a)^2]

for a > 0, b > 0, y >= 0. Here, b is the scale parameter scale, and a is a shape parameter. The cumulative distribution function is

F(y) = 1 - [1 + (y/b)^a]^(-1) = [1 + (y/b)^(-a)]^(-1).

The mean is

E(Y) = b gamma(1 + 1/a) gamma(1 - 1/a)

provided a > 1; these are returned as the fitted values. This family function handles multiple responses.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Note

See the notes in genbetaII.

Author(s)

T. W. Yee

References

Kleiber, C. and Kotz, S. (2003) Statistical Size Distributions in Economics and Actuarial Sciences, Hoboken, NJ, USA: Wiley-Interscience.

See Also

Fisk, genbetaII, betaII, dagum, sinmad, inv.lomax, lomax, paralogistic, inv.paralogistic, simulate.vlm.

Examples

fdata <- data.frame(y = rfisk(n = 200, shape = exp(1), scale = exp(2)))
fit <- vglm(y ~ 1, fisk(lss = FALSE), data = fdata, trace = TRUE)
fit <- vglm(y ~ 1, fisk(ishape1.a = exp(2)), data = fdata, trace = TRUE)
coef(fit, matrix = TRUE)
Coef(fit)
summary(fit)

[Package VGAM version 1.0-6 Index]