flourbeetle {VGAM}	R Documentation

Mortality of Flour Beetles from Carbon Disulphide

Description

The flourbeetle data frame has 8 rows and 4 columns. Two columns are explanatory, the other two are responses.

Usage

data(flourbeetle)

Format

This data frame contains the following columns:

logdose

log10 applied to CS2mgL.

CS2mgL

a numeric vector, the concentration of gaseous carbon disulphide in mg per litre.

exposed

a numeric vector, counts; the number of beetles exposed to the poison.

killed

a numeric vector, counts; the numbers killed.

Details

These data were originally given in Table IV of Bliss (1935) and are the combination of two series of toxicological experiments involving Tribolium confusum, also known as the flour beetle. Groups of such adult beetles were exposed for 5 hours of gaseous carbon disulphide at different concentrations, and their mortality measured.

Source

Bliss, C.I., 1935. The calculation of the dosage-mortality curve. Annals of Applied Biology, 22, 134–167.

See Also

binomialff, probit.

Examples

fit1 <- vglm(cbind(killed, exposed - killed) ~ logdose, binomialff(link = probit),
             data = flourbeetle, trace = TRUE)
summary(fit1)

[Package VGAM version 1.0-6 Index]