GenbetaII {VGAM}	R Documentation

The Generalized Beta II Distribution

Description

Density for the generalized beta II distribution with shape parameters a and p and q, and scale parameter scale.

Usage

dgenbetaII(x, scale = 1, shape1.a, shape2.p, shape3.q, log = FALSE)

Arguments

x
vector of quantiles.

shape1.a, shape2.p, shape3.q
positive shape parameters.

scale
positive scale parameter.

log
Logical. If log = TRUE then the logarithm of the density is returned.

Details

See genbetaII, which is the VGAM family function for estimating the parameters by maximum likelihood estimation. Several distributions, such as the Singh-Maddala, are special case of this flexible 4-parameter distribution. The product of shape1.a and shape2.p determines the behaviour of the density at the origin.

Value

dgenbetaII gives the density.

Author(s)

T. W. Yee

References

Kleiber, C. and Kotz, S. (2003) Statistical Size Distributions in Economics and Actuarial Sciences, Hoboken, NJ, USA: Wiley-Interscience.

See Also

genbetaII.

Examples

dgenbetaII(0, shape1.a = 1/4, shape2.p = 4, shape3.q = 3)
dgenbetaII(0, shape1.a = 1/4, shape2.p = 2, shape3.q = 3)
dgenbetaII(0, shape1.a = 1/4, shape2.p = 8, shape3.q = 3)

[Package VGAM version 1.0-6 Index]