gew {VGAM}	R Documentation

General Electric and Westinghouse Data

Description

General Electric and Westinghouse capital data.

Usage

data(gew)

Format

A data frame with 20 observations on the following 7 variables. All variables are numeric vectors. Variables ending in .g correspond to General Electric and those ending in .w are Westinghouse.

year

The observations are the years from 1934 to 1953

invest.g, invest.w

investment figures. These are I= Gross investment = additions to plant and equipment plus maintenance and repairs in millions of dollars deflated by P_1.

capital.g, capital.w

capital stocks. These are C= The stock of plant and equipment = accumulated sum of net additions to plant and equipment deflated by P_1 minus depreciation allowance deflated by P_3.

value.g, value.w

market values. These are F= Value of the firm = price of common and preferred shares at December 31 (or average price of December 31 and January 31 of the following year) times number of common and preferred shares outstanding plus total book value of debt at December 31 in millions of dollars deflated by P_2.

Details

These data are a subset of a table in Boot and de Wit (1960), also known as the Grunfeld data. It is used a lot in econometrics, e.g., for seemingly unrelated regressions (see SURff).

Here, P_1 = Implicit price deflator of producers durable equipment (base 1947), P_2 = Implicit price deflator of G.N.P. (base 1947), P_3 = Depreciation expense deflator = ten years moving average of wholesale price index of metals and metal products (base 1947).

Source

Table 10 of: Boot, J. C. G. and de Wit, G. M. (1960) Investment Demand: An Empirical Contribution to the Aggregation Problem. International Economic Review, 1, 3–30.

Grunfeld, Y. (1958) The Determinants of Corporate Investment. Unpublished PhD Thesis (Chicago).

References

Zellner, A. (1962) An efficient method of estimating seemingly unrelated regressions and tests for aggregation bias. Journal of the American Statistical Association, 57, 348–368.

See Also

SURff, http://statmath.wu.ac.at/~zeileis/grunfeld.

Examples

str(gew)

[Package VGAM version 1.0-6 Index]