grain.us {VGAM}	R Documentation

Grain Prices Data in USA

Description

A 4-column matrix.

Usage

data(grain.us)

Format

The columns are:

wheat.flour

numeric

corn

numeric

wheat

numeric

rye

numeric

Details

Monthly averages of grain prices in the United States for wheat flour, corn, wheat, and rye for the period January 1961 through October 1972. The units are US dollars per 100 pound sack for wheat flour, and per bushel for corn, wheat and rye.

Source

Ahn and Reinsel (1988).

References

Ahn, S. K and Reinsel, G. C. (1988) Nested reduced-rank autoregressive models for multiple time series. Journal of the American Statistical Association, 83, 849–856.

Examples

## Not run: 
cgrain <- scale(grain.us, scale = FALSE)  # Center the time series only
fit <- vglm(cgrain ~ 1, rrar(Rank = c(4, 1)),
            epsilon = 1e-3, stepsize = 0.5, trace = TRUE, maxit = 50)
summary(fit)

## End(Not run)

[Package VGAM version 1.0-6 Index]