hspider {VGAM}	R Documentation

Hunting Spider Data

Description

Abundance of hunting spiders in a Dutch dune area.

Usage

data(hspider)

Format

A data frame with 28 observations (sites) on the following 18 variables.

WaterCon

Log percentage of soil dry mass.

BareSand

Log percentage cover of bare sand.

FallTwig

Log percentage cover of fallen leaves and twigs.

CoveMoss

Log percentage cover of the moss layer.

CoveHerb

Log percentage cover of the herb layer.

ReflLux

Reflection of the soil surface with cloudless sky.

Alopacce

Abundance of Alopecosa accentuata.

Alopcune

Abundance of Alopecosa cuneata.

Alopfabr

Abundance of Alopecosa fabrilis.

Arctlute

Abundance of Arctosa lutetiana.

Arctperi

Abundance of Arctosa perita.

Auloalbi

Abundance of Aulonia albimana.

Pardlugu

Abundance of Pardosa lugubris.

Pardmont

Abundance of Pardosa monticola.

Pardnigr

Abundance of Pardosa nigriceps.

Pardpull

Abundance of Pardosa pullata.

Trocterr

Abundance of Trochosa terricola.

Zoraspin

Abundance of Zora spinimana.

Details

The data, which originally came from Van der Aart and Smeek-Enserink (1975) consists of abundances (numbers trapped over a 60 week period) and 6 environmental variables. There were 28 sites.

This data set has been often used to illustrate ordination, e.g., using canonical correspondence analysis (CCA). In the example below, the data is used for constrained quadratic ordination (CQO; formerly called canonical Gaussian ordination or CGO), a numerically intensive method that has many superior qualities. See cqo for details.

References

Van der Aart, P. J. M. and Smeek-Enserink, N. (1975) Correlations between distributions of hunting spiders (Lycosidae, Ctenidae) and environmental characteristics in a dune area. Netherlands Journal of Zoology, 25, 1–45.

Examples

summary(hspider)

## Not run: 
# Standardize the environmental variables:
hspider[, 1:6] <- scale(subset(hspider, select = WaterCon:ReflLux))

# Fit a rank-1 binomial CAO
hsbin <- hspider  # Binary species data
hsbin[, -(1:6)] <- as.numeric(hsbin[, -(1:6)] > 0)
set.seed(123)
ahsb1 <- cao(cbind(Alopcune, Arctlute, Auloalbi, Zoraspin) ~
             WaterCon + ReflLux,
             family = binomialff(multiple.responses = TRUE),
             df1.nl = 2.2, Bestof = 3, data = hsbin)
par(mfrow = 2:1, las = 1)
lvplot(ahsb1, type = "predictors", llwd = 2, ylab = "logit p", lcol = 1:9)
persp(ahsb1, rug = TRUE, col = 1:10, lwd = 2)
coef(ahsb1)

## End(Not run)

[Package VGAM version 1.0-6 Index]