hunua {VGAM}	R Documentation

Hunua Ranges Data

Description

The hunua data frame has 392 rows and 18 columns. Altitude is explanatory, and there are binary responses (presence/absence = 1/0 respectively) for 17 plant species.

Usage

data(hunua)

Format

This data frame contains the following columns:

agaaus

Agathis australis, or Kauri

beitaw

Beilschmiedia tawa, or Tawa

corlae

Corynocarpus laevigatus

cyadea

Cyathea dealbata

cyamed

Cyathea medullaris

daccup

Dacrydium cupressinum

dacdac

Dacrycarpus dacrydioides

eladen

Elaecarpus dentatus

hedarb

Hedycarya arborea

hohpop

Species name unknown

kniexc

Knightia excelsa, or Rewarewa

kuneri

Kunzea ericoides

lepsco

Leptospermum scoparium

metrob

Metrosideros robusta

neslan

Nestegis lanceolata

rhosap

Rhopalostylis sapida

vitluc

Vitex lucens, or Puriri

altitude

meters above sea level

Details

These were collected from the Hunua Ranges, a small forest in southern Auckland, New Zealand. At 392 sites in the forest, the presence/absence of 17 plant species was recorded, as well as the altitude. Each site was of area size 200m^2.

Source

Dr Neil Mitchell, University of Auckland.

See Also

waitakere.

Examples

# Fit a GAM using vgam() and compare it with the Waitakere Ranges one
fit.h <- vgam(agaaus ~ s(altitude, df = 2), binomialff, data = hunua)
## Not run: 
plot(fit.h, se = TRUE, lcol = "orange", scol = "orange",
     llwd = 2, slwd = 2, main = "Orange is Hunua, Blue is Waitakere") 
## End(Not run)
head(predict(fit.h, hunua, type = "response"))

fit.w <- vgam(agaaus ~ s(altitude, df = 2), binomialff, data = waitakere)
## Not run: 
plot(fit.w, se = TRUE, lcol = "blue", scol = "blue", add = TRUE) 
## End(Not run)
head(predict(fit.w, hunua, type = "response"))   # Same as above?

[Package VGAM version 1.0-6 Index]