identitylink {VGAM}	R Documentation

Identity Link Function

Description

Computes the identity transformation, including its inverse and the first two derivatives.

Usage

identitylink(theta, inverse = FALSE, deriv = 0, short = TRUE, tag = FALSE)
 negidentity(theta, inverse = FALSE, deriv = 0, short = TRUE, tag = FALSE)

Arguments

theta
Numeric or character. See below for further details.

inverse, deriv, short, tag
Details at Links.

Details

The identity link function g(theta)=theta should be available to every parameter estimated by the VGAM library. However, it usually results in numerical problems because the estimates lie outside the permitted range. Consequently, the result may contain Inf, -Inf, NA or NaN.

The function negidentity is the negative-identity link function and corresponds to g(theta)=-theta. This is useful for some models, e.g., in the literature supporting the gevff function it seems that half of the authors use xi=-k for the shape parameter and the other half use k instead of xi.

Value

For identitylink(): for deriv = 0, the identity of theta, i.e., theta when inverse = FALSE, and if inverse = TRUE then theta. For deriv = 1, then the function returns d eta / d theta as a function of theta if inverse = FALSE, else if inverse = TRUE then it returns the reciprocal.

For negidentity(): the results are similar to identitylink() except for a sign change in most cases.

Author(s)

Thomas W. Yee

References

McCullagh, P. and Nelder, J. A. (1989) Generalized Linear Models, 2nd ed. London: Chapman & Hall.

See Also

Links, loge, logit, probit, powerlink.

Examples

identitylink((-5):5)
identitylink((-5):5, deriv = 1)
identitylink((-5):5, deriv = 2)
negidentity((-5):5)
negidentity((-5):5, deriv = 1)
negidentity((-5):5, deriv = 2)

[Package VGAM version 1.0-6 Index]