inv.lomax {VGAM}	R Documentation

Inverse Lomax Distribution Family Function

Description

Maximum likelihood estimation of the 2-parameter inverse Lomax distribution.

Usage

inv.lomax(lscale = "loge", lshape2.p = "loge", iscale = NULL,
    ishape2.p = NULL, imethod = 1, gscale = exp(-5:5),
    gshape2.p = exp(-5:5), probs.y = c(0.25, 0.5, 0.75), zero = "shape2.p")

Arguments

lscale, lshape2.p
Parameter link functions applied to the (positive) parameters b, and p. See Links for more choices.

iscale, ishape2.p, imethod, zero
See CommonVGAMffArguments for information. For imethod = 2 a good initial value for ishape2.p is needed to obtain a good estimate for the other parameter.

gscale, gshape2.p
See CommonVGAMffArguments for information.

probs.y
See CommonVGAMffArguments for information.

Details

The 2-parameter inverse Lomax distribution is the 4-parameter generalized beta II distribution with shape parameters a=q=1. It is also the 3-parameter Dagum distribution with shape parameter a=1, as well as the beta distribution of the second kind with q=1. More details can be found in Kleiber and Kotz (2003).

The inverse Lomax distribution has density

f(y) = p y^(p-1) / [b^p (1 + y/b)^(p+1)]

for b > 0, p > 0, y >= 0. Here, b is the scale parameter scale, and p is a shape parameter. The mean does not seem to exist; the median is returned as the fitted values. This family function handles multiple responses.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Note

See the notes in genbetaII.

Author(s)

T. W. Yee

References

Kleiber, C. and Kotz, S. (2003) Statistical Size Distributions in Economics and Actuarial Sciences, Hoboken, NJ, USA: Wiley-Interscience.

See Also

inv.lomax, genbetaII, betaII, dagum, sinmad, fisk, lomax, paralogistic, inv.paralogistic, simulate.vlm.

Examples

idata <- data.frame(y = rinv.lomax(n = 2000, scale = exp(2), exp(1)))
fit <- vglm(y ~ 1, inv.lomax, data = idata, trace = TRUE)
fit <- vglm(y ~ 1, inv.lomax(iscale = exp(3)), data = idata,
            trace = TRUE, epsilon = 1e-8, crit = "coef")
coef(fit, matrix = TRUE)
Coef(fit)
summary(fit)

[Package VGAM version 1.0-6 Index]