Inv.paralogistic {VGAM}	R Documentation

The Inverse Paralogistic Distribution

Description

Density, distribution function, quantile function and random generation for the inverse paralogistic distribution with shape parameters a and p, and scale parameter scale.

Usage

dinv.paralogistic(x, scale = 1, shape1.a, log = FALSE)
pinv.paralogistic(q, scale = 1, shape1.a, lower.tail = TRUE, log.p = FALSE)
qinv.paralogistic(p, scale = 1, shape1.a, lower.tail = TRUE, log.p = FALSE)
rinv.paralogistic(n, scale = 1, shape1.a)

Arguments

x, q
vector of quantiles.

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

shape1.a
shape parameter.

scale
scale parameter.

log
Logical. If log = TRUE then the logarithm of the density is returned.

lower.tail, log.p
Same meaning as in pnorm or qnorm.

Details

See inv.paralogistic, which is the VGAM family function for estimating the parameters by maximum likelihood estimation.

Value

dinv.paralogistic gives the density, pinv.paralogistic gives the distribution function, qinv.paralogistic gives the quantile function, and rinv.paralogistic generates random deviates.

Note

The inverse paralogistic distribution is a special case of the 4-parameter generalized beta II distribution.

Author(s)

T. W. Yee

References

Kleiber, C. and Kotz, S. (2003) Statistical Size Distributions in Economics and Actuarial Sciences, Hoboken, NJ, USA: Wiley-Interscience.

See Also

inv.paralogistic, genbetaII.

Examples

idata <- data.frame(y = rinv.paralogistic(n = 3000, exp(1), scale = exp(2)))
fit <- vglm(y ~ 1, inv.paralogistic(lss = FALSE, ishape1.a = 2.1),
            data = idata, trace = TRUE, crit = "coef")
coef(fit, matrix = TRUE)
Coef(fit)

[Package VGAM version 1.0-6 Index]