is.parallel {VGAM}	R Documentation

Parallelism Constraint Matrices

Description

Returns a logical vector from a test of whether an object such as a matrix or VGLM object corresponds to a parallelism assumption.

Usage

is.parallel.matrix(object, ...)
is.parallel.vglm(object, type = c("term", "lm"), ...)

Arguments

object
an object such as a constraint matrix or a vglm object.

type
passed into constraints.

...
additional optional arguments. Currently unused.

Details

These functions may be useful for categorical models such as propodds, cumulative, acat, cratio, sratio, multinomial.

Value

A vector of logicals, testing whether each constraint matrix is a one-column matrix of ones. Note that parallelism can still be thought of as holding if the constraint matrix has a non-zero but constant values, however, this is currently not implemented. No checking is done that the constraint matrices have the same number of rows.

See Also

constraints, vglm.

Examples

## Not run:  require("VGAMdata")
fit <- vglm(educ ~ sm.bs(age) * sex + ethnicity,
            cumulative(parallel = TRUE), head(xs.nz, 200))
is.parallel(fit)
is.parallel(fit, type = "lm")  # For each column of the LM matrix

## End(Not run)

[Package VGAM version 1.0-6 Index]