is.zero {VGAM}	R Documentation

Zero Constraint Matrices

Description

Returns a logical vector from a test of whether an object such as a matrix or VGLM object corresponds to a 'zero' assumption.

Usage

is.zero.matrix(object, ...)
is.zero.vglm(object, ...)

Arguments

object
an object such as a coefficient matrix of a vglm object, or a vglm object.

...
additional optional arguments. Currently unused.

Details

These functions test the effect of the zero argument on a vglm object or the coefficient matrix of a vglm object. The latter is obtained by coef(vglmObject, matrix = TRUE).

Value

A vector of logicals, testing whether each linear/additive predictor has the zero argument applied to it. It is TRUE if that linear/additive predictor is intercept-only, i.e., all other regression coefficients are set to zero.

No checking is done for the intercept term at all, i.e., that it was estimated in the first place.

See Also

constraints, vglm.

Examples

coalminers <- transform(coalminers, Age = (age - 42) / 5)
fit <- vglm(cbind(nBnW,nBW,BnW,BW) ~ Age, binom2.or(zero = NULL), coalminers)
is.zero(fit)
is.zero(coef(fit, matrix = TRUE))

[Package VGAM version 1.0-6 Index]