lakeO {VGAM}	R Documentation

Annual catches on Lake Otamangakau from October 1974 to October 1989

Description

Rainbow and brown trout catches by a Mr Swainson at Lake Otamangakau in the central North Island of New Zealand during the 1970s and 1980s.

Usage

data(lakeO)

Format

A data frame with 15 observations on the following 5 variables.

year

a numeric vector, the season began on 1 October of the year and ended 12 months later.

total.fish

a numeric vector, the total number of fish caught during the season. Simply the sum of brown and rainbow trout.

brown

a numeric vector, the number of brown trout (Salmo trutta) caught.

rainbow

a numeric vector, the number of rainbow trout (Oncorhynchus mykiss) caught.

visits

a numeric vector, the number of visits during the season that the angler made to the lake. It is necessary to assume that the visits were of an equal time length in order to interpret the usual Poisson regressions.

Details

The data was extracted from the season summaries at Lake Otamangakau by Anthony Swainson for the seasons 1974–75 to 1988–89. Mr Swainson was one of a small group of regular fly fishing anglers and kept a diary of his catches. Lake Otamangakau is a lake of area 1.8 squared km and has a maximum depth of about 12m, and is located in the central North Island of New Zealand. It is trout-infested and known for its trophy-sized fish.

See also trapO.

Source

Table 7.2 of the reference below. Thanks to Dr Michel Dedual for a copy of the report and for help reading the final year's data. The report is available from TWY on request.

References

Dedual, M. and MacLean, G. and Rowe, D. and Cudby, E., The Trout Population and Fishery of Lake Otamangakau—Interim Report. National Institute of Water and Atmospheric Research, Hamilton, New Zealand. Consultancy Report Project No. ELE70207, (Dec 1996).

Examples

data(lakeO)
lakeO
summary(lakeO)

[Package VGAM version 1.0-6 Index]