lgamma1 {VGAM}	R Documentation

Log-gamma Distribution Family Function

Description

Estimation of the parameter of the standard and nonstandard log-gamma distribution.

Usage

lgamma1(lshape = "loge", ishape = NULL)
lgamma3(llocation = "identitylink", lscale = "loge", lshape = "loge",
        ilocation = NULL, iscale = NULL, ishape = 1,
        zero = c("scale", "shape"))

Arguments

llocation, lscale
Parameter link function applied to the location parameter a and the positive scale parameter b. See Links for more choices.

lshape
Parameter link function applied to the positive shape parameter k. See Links for more choices.

ishape
Initial value for k. If given, it must be positive. If failure to converge occurs, try some other value. The default means an initial value is determined internally.

ilocation, iscale
Initial value for a and b. The defaults mean an initial value is determined internally for each.

zero
An integer-valued vector specifying which linear/additive predictors are modelled as intercepts only. The values must be from the set {1,2,3}. The default value means none are modelled as intercept-only terms. See CommonVGAMffArguments for more information.

Details

The probability density function of the standard log-gamma distribution is given by

f(y;k) = exp[ky - exp(y)]/gamma(k),

for parameter k>0 and all real y. The mean of Y is digamma(k) (returned as the fitted values) and its variance is trigamma(k).

For the non-standard log-gamma distribution, one replaces y by (y-a)/b, where a is the location parameter and b is the positive scale parameter. Then the density function is

f(y) = exp[k(y-a)/b - exp((y-a)/b)]/(b*gamma(k)).

The mean and variance of Y are a + b*digamma(k) (returned as the fitted values) and b^2 * trigamma(k), respectively.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Note

The standard log-gamma distribution can be viewed as a generalization of the standard type 1 extreme value density: when k = 1 the distribution of -Y is the standard type 1 extreme value distribution.

The standard log-gamma distribution is fitted with lgamma1 and the non-standard (3-parameter) log-gamma distribution is fitted with lgamma3.

Author(s)

T. W. Yee

References

Kotz, S. and Nadarajah, S. (2000) Extreme Value Distributions: Theory and Applications, pages 48–49, London: Imperial College Press.

Johnson, N. L. and Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, 2nd edition, Volume 2, p.89, New York: Wiley.

See Also

rlgamma, gengamma.stacy, prentice74, gamma1, lgamma.

Examples

ldata <- data.frame(y = rlgamma(100, shape = exp(1)))
fit <- vglm(y ~ 1, lgamma1, data = ldata, trace = TRUE, crit = "coef")
summary(fit)
coef(fit, matrix = TRUE)
Coef(fit)

ldata <- data.frame(x2 = runif(nn <- 5000))  # Another example
ldata <- transform(ldata, loc = -1 + 2 * x2, Scale = exp(1))
ldata <- transform(ldata, y = rlgamma(nn, loc, scale = Scale, shape = exp(0)))
fit2 <- vglm(y ~ x2, lgamma3, data = ldata, trace = TRUE, crit = "c")
coef(fit2, matrix = TRUE)

[Package VGAM version 1.0-6 Index]