Lindley {VGAM}	R Documentation

The Lindley Distribution

Description

Density, cumulative distribution function, and random generation for the Lindley distribution.

Usage

dlind(x, theta, log = FALSE)
plind(q, theta, lower.tail = TRUE, log.p = FALSE)
rlind(n, theta)

Arguments

x, q
vector of quantiles.

n
number of observations. Same as in runif.

log
Logical. If log = TRUE then the logarithm of the density is returned.

theta
positive parameter.

lower.tail, log.p
Same meaning as in pnorm or qnorm.

Details

See lindley for details.

Value

dlind gives the density, plind gives the cumulative distribution function, and rlind generates random deviates.

Author(s)

T. W. Yee and Kai Huang

See Also

lindley.

Examples

theta <- exp(-1); x <- seq(0.0, 17, length = 700)
dlind(0:10, theta)
## Not run: 
plot(x, dlind(x, theta), type = "l", las = 1, col = "blue",
     main = "dlind(x, theta = exp(-1))")
abline(h = 1, col = "grey", lty = "dashed") 
## End(Not run)

[Package VGAM version 1.0-6 Index]