lindley {VGAM}	R Documentation

1-parameter Gamma Distribution

Description

Estimates the (1-parameter) Lindley distribution by maximum likelihood estimation.

Usage

lindley(link = "loge", itheta = NULL, zero = NULL)

Arguments

link
Link function applied to the (positive) parameter. See Links for more choices.

itheta, zero
See CommonVGAMffArguments for information.

Details

The density function is given by

f(y; theta) = theta^2 * (1 + y) * exp(-theta * y) / (1 + theta)

for theta > 0 and y > 0. The mean of Y (returned as the fitted values) is mu = (theta + 2) / (theta * (theta + 1)). The variance is (theta^2 + 4 * theta + 2) / (theta * (theta + 1))^2.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Note

This VGAM family function can handle multiple responses (inputted as a matrix). Fisher scoring is implemented.

Author(s)

T. W. Yee

References

Lindley, D. V. (1958) Fiducial distributions and Bayes' theorem. Journal of the Royal Statistical Society, Series B, Methodological, 20, 102–107.

Ghitany, M. E. and Atieh, B. and Nadarajah, S. (2008) Lindley distribution and its application. Math. Comput. Simul., 78, 493–506.

See Also

dlind, gammaR, simulate.vlm.

Examples

ldata <- data.frame(y = rlind(n = 1000, theta = exp(3)))
fit <- vglm(y ~ 1, lindley, data = ldata, trace = TRUE, crit = "coef")
coef(fit, matrix = TRUE)
Coef(fit)
summary(fit)

[Package VGAM version 1.0-6 Index]