linkfun {VGAM}	R Documentation

Link Functions

Description

Generic function for returning the link functions of a fitted object.

Usage

linkfun(object, ...)

Arguments

object
An object which has parameter link functions.

...
Other arguments fed into the specific methods function of the model.

Details

Fitted models in the VGAM have parameter link functions. This generic function returns these.

Value

The value returned depends specifically on the methods function invoked.

Author(s)

Thomas W. Yee

See Also

linkfun.vglm, multilogit, vglm.

Examples

pneumo <- transform(pneumo, let = log(exposure.time))
fit1 <- vglm(cbind(normal, mild, severe) ~ let, propodds, data = pneumo)
coef(fit1, matrix = TRUE)
linkfun(fit1)
linkfun(fit1, earg = TRUE)

fit2 <- vglm(cbind(normal, mild, severe) ~ let, multinomial, data = pneumo)
coef(fit2, matrix = TRUE)
linkfun(fit2)
linkfun(fit2, earg = TRUE)

[Package VGAM version 1.0-6 Index]