linkfun.vglm {VGAM}	R Documentation

Link Functions for VGLMs

Description

Returns the link functions, and parameter names, for vector generalized linear models (VGLMs).

Usage

linkfun.vglm(object, earg = FALSE, ...)

Arguments

object
Object of class "vglm", i.e., a VGLM object.

earg
Logical. Return the extra arguments associated with each link function? If TRUE then a list is returned.

...
Arguments that might be used in the future.

Details

All fitted VGLMs have a link function applied to each parameter. This function returns these, and optionally, the extra arguments associated with them.

Value

Usually just a (named) character string, with the link functions in order. It is named with the parameter names. If earg = TRUE then a list with the following components.

link
The default output.

earg
The extra arguments, in order.

Note

Presently, the multinomial logit model has only one link function, multilogit, so a warning is not issued for that link. For other models, if the number of link functions does not equal M then a warning may be issued.

Author(s)

Thomas W. Yee

See Also

linkfun, multilogit, vglm.

Examples

fit1 <- vgam(cbind(r1, r2) ~ s(year, df = 3), gev(zero = 2:3), data = venice)
coef(fit1, matrix = TRUE)
linkfun(fit1)
linkfun(fit1, earg = TRUE)

[Package VGAM version 1.0-6 Index]