logF {VGAM}	R Documentation

Natural Exponential Family Generalized Hyperbolic Secant Distribution Family Function

Description

Maximum likelihood estimation of the 1-parameter log F distribution.

Usage

 logF(lshape1 = "loge", lshape2 = "loge",
      ishape1 = NULL, ishape2 = 1, imethod = 1)

Arguments

lshape1, lshape2
Parameter link functions for the shape parameters. Called alpha and beta respectively. See Links for more choices.

ishape1, ishape2
Optional initial values for the shape parameters. If given, it must be numeric and values are recycled to the appropriate length. The default is to choose the value internally. See CommonVGAMffArguments for more information.

imethod
Initialization method. Either the value 1, 2, or .... See CommonVGAMffArguments for more information.

Details

The density for this distribution is

f(y; alpha, beta) = exp(α y) / [B(α,β) * (1 + exp(y))^(α + β)]

where y is real, α > 0, β > 0, B(., .) is the beta function beta.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm and vgam.

Author(s)

Thomas W. Yee

References

Jones, M. C. (2008). On a class of distributions with simple exponential tails. Statistica Sinica, 18(3), 1101–1110.

See Also

dlogF, logff.

Examples

nn <- 1000
ldata <- data.frame(y1 = rnorm(nn, m = +1, sd = exp(2)),  # Not proper data
                    x2 = rnorm(nn, m = -1, sd = exp(2)),
                    y2 = rnorm(nn, m = -1, sd = exp(2)))  # Not proper data
fit1 <- vglm(y1 ~ 1 , logF, data = ldata, trace = TRUE)
fit2 <- vglm(y2 ~ x2, logF, data = ldata, trace = TRUE)
coef(fit2, matrix = TRUE)
summary(fit2)
vcov(fit2)

head(fitted(fit1))
with(ldata, mean(y1))
max(abs(head(fitted(fit1)) - with(ldata, mean(y1))))

[Package VGAM version 1.0-6 Index]