logc {VGAM}	R Documentation

Complementary-log Link Function

Description

Computes the Complementary-log Transformation, Including its Inverse and the First Two Derivatives.

Usage

logc(theta, bvalue = NULL, inverse = FALSE, deriv = 0,
     short = TRUE, tag = FALSE)

Arguments

theta
Numeric or character. See below for further details.

bvalue
See Links.

inverse, deriv, short, tag
Details at Links.

Details

The complementary-log link function is suitable for parameters that are less than unity. Numerical values of theta close to 1 or out of range result in Inf, -Inf, NA or NaN.

Value

For deriv = 0, the log of theta, i.e., log(1-theta) when inverse = FALSE, and if inverse = TRUE then 1-exp(theta).

For deriv = 1, then the function returns d eta / d theta as a function of theta if inverse = FALSE, else if inverse = TRUE then it returns the reciprocal.

Here, all logarithms are natural logarithms, i.e., to base e.

Note

Numerical instability may occur when theta is close to 1. One way of overcoming this is to use bvalue.

Author(s)

Thomas W. Yee

References

McCullagh, P. and Nelder, J. A. (1989) Generalized Linear Models, 2nd ed. London: Chapman & Hall.

See Also

Links, loge, cloglog, loglog, logoff.

Examples

## Not run: 
logc(seq(-0.2, 1.1, by = 0.1))  # Has NAs

## End(Not run)
logc(seq(-0.2, 1.1, by = 0.1), bvalue = 1 - .Machine$double.eps)  # Has no NAs

[Package VGAM version 1.0-6 Index]