logistic {VGAM}	R Documentation

Logistic Distribution Family Function

Description

Estimates the location and scale parameters of the logistic distribution by maximum likelihood estimation.

Usage

logistic1(llocation = "identitylink", scale.arg = 1, imethod = 1)
logistic(llocation = "identitylink", lscale = "loge",
         ilocation = NULL, iscale = NULL, imethod = 1, zero = "scale")

Arguments

llocation, lscale
Parameter link functions applied to the location parameter l and scale parameter s. See Links for more choices, and CommonVGAMffArguments for more information.

scale.arg
Known positive scale parameter (called s below).

ilocation, iscale
See CommonVGAMffArguments for information.

imethod, zero
See CommonVGAMffArguments for information.

Details

The two-parameter logistic distribution has a density that can be written as

f(y;l,s) = exp[-(y-l)/s] / [s * ( 1 + exp[-(y-l)/s] )^2]

where s > 0 is the scale parameter, and l is the location parameter. The response -Inf<y<Inf. The mean of Y (which is the fitted value) is l and its variance is pi^2 s^2 / 3.

A logistic distribution with scale = 0.65 (see dlogis) resembles dt with df = 7; see logistic1 and studentt.

logistic1 estimates the location parameter only while logistic estimates both parameters. By default, eta1 = l and eta2 = log(s) for logistic.

logistic can handle multiple responses.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, rrvglm and vgam.

Note

Fisher scoring is used, and the Fisher information matrix is diagonal.

Author(s)

T. W. Yee

References

Johnson, N. L. and Kotz, S. and Balakrishnan, N. (1994) Continuous Univariate Distributions, 2nd edition, Volume 1, New York: Wiley. Chapter 15.

Forbes, C., Evans, M., Hastings, N. and Peacock, B. (2011) Statistical Distributions, Hoboken, NJ, USA: John Wiley and Sons, Fourth edition.

Castillo, E., Hadi, A. S., Balakrishnan, N. Sarabia, J. S. (2005) Extreme Value and Related Models with Applications in Engineering and Science, Hoboken, NJ, USA: Wiley-Interscience, p.130.

deCani, J. S. and Stine, R. A. (1986) A Note on Deriving the Information Matrix for a Logistic Distribution, The American Statistician, 40, 220–222.

See Also

rlogis, CommonVGAMffArguments, logit, cumulative, bilogistic, simulate.vlm.

Examples

# Location unknown, scale known
ldata <- data.frame(x2 = runif(nn <- 500))
ldata <- transform(ldata, y1 = rlogis(nn, loc = 1 + 5*x2, scale = exp(2)))
fit1 <- vglm(y1 ~ x2, logistic1(scale = exp(2)), data = ldata, trace = TRUE)
coef(fit1, matrix = TRUE)

# Both location and scale unknown
ldata <- transform(ldata, y2 = rlogis(nn, loc = 1 + 5*x2, scale = exp(0 + 1*x2)))
fit2 <- vglm(cbind(y1, y2) ~ x2, logistic, data = ldata, trace = TRUE)
coef(fit2, matrix = TRUE)
vcov(fit2)
summary(fit2)

[Package VGAM version 1.0-6 Index]