logit {VGAM}	R Documentation

Logit Link Function

Description

Computes the logit transformation, including its inverse and the first two derivatives.

Usage

logit(theta, bvalue = NULL, inverse = FALSE, deriv = 0,
      short = TRUE, tag = FALSE)
extlogit(theta, min = 0, max = 1, bminvalue = NULL, bmaxvalue = NULL,
         inverse = FALSE, deriv = 0, short = TRUE, tag = FALSE)

Arguments

theta
Numeric or character. See below for further details.

bvalue, bminvalue, bmaxvalue
See Links. These are boundary values. For extlogit, values of theta less than or equal to A or greater than or equal to B can be replaced by bminvalue and bmaxvalue.

min, max
For extlogit, min gives A, max gives B, and for out of range values, bminvalue and bmaxvalue.

inverse, deriv, short, tag
Details at Links.

Details

The logit link function is very commonly used for parameters that lie in the unit interval. It is the inverse CDF of the logistic distribution. Numerical values of theta close to 0 or 1 or out of range result in Inf, -Inf, NA or NaN.

The extended logit link function extlogit should be used more generally for parameters that lie in the interval (A,B), say. The formula is

log((theta-A)/(B-theta))

and the default values for A and B correspond to the ordinary logit function. Numerical values of theta close to A or B or out of range result in Inf, -Inf, NA or NaN. However these can be replaced by values bminvalue and bmaxvalue first before computing the link function.

Value

For logit with deriv = 0, the logit of theta, i.e., log(theta/(1-theta)) when inverse = FALSE, and if inverse = TRUE then exp(theta)/(1+exp(theta)).

For deriv = 1, then the function returns d eta / d theta as a function of theta if inverse = FALSE, else if inverse = TRUE then it returns the reciprocal.

Here, all logarithms are natural logarithms, i.e., to base e.

Note

Numerical instability may occur when theta is close to 1 or 0 (for logit), or close to A or B for extlogit. One way of overcoming this is to use, e.g., bvalue.

In terms of the threshold approach with cumulative probabilities for an ordinal response this link function corresponds to the univariate logistic distribution (see logistic).

Author(s)

Thomas W. Yee

References

McCullagh, P. and Nelder, J. A. (1989) Generalized Linear Models, 2nd ed. London: Chapman & Hall.

See Also

Links, logitoffsetlink, probit, cloglog, cauchit, logistic1, loge, Logistic, multilogit.

Examples

p <- seq(0.01, 0.99, by = 0.01)
logit(p)
max(abs(logit(logit(p), inverse = TRUE) - p))  # Should be 0

p <- c(seq(-0.02, 0.02, by = 0.01), seq(0.97, 1.02, by = 0.01))
logit(p)  # Has NAs
logit(p, bvalue = .Machine$double.eps)  # Has no NAs

p <- seq(0.9, 2.2, by = 0.1)
extlogit(p, min = 1, max = 2,
          bminvalue = 1 + .Machine$double.eps,
          bmaxvalue = 2 - .Machine$double.eps)  # Has no NAs

## Not run:  par(mfrow = c(2,2), lwd = (mylwd <- 2))
y <- seq(-4, 4, length = 100)
p <- seq(0.01, 0.99, by = 0.01)
for (d in 0:1) {
  myinv <- (d > 0)
  matplot(p, cbind( logit(p, deriv = d, inverse = myinv),
                   probit(p, deriv = d, inverse = myinv)),
          type = "n", col = "purple", ylab = "transformation", las = 1,
          main = if (d ==  0) "Some probability link functions"
          else "1 / first derivative")
  lines(p,   logit(p, deriv = d, inverse = myinv), col = "limegreen")
  lines(p,  probit(p, deriv = d, inverse = myinv), col = "purple")
  lines(p, cloglog(p, deriv = d, inverse = myinv), col = "chocolate")
  lines(p, cauchit(p, deriv = d, inverse = myinv), col = "tan")
  if (d ==  0) {
    abline(v = 0.5, h = 0, lty = "dashed")
    legend(0, 4.5, c("logit", "probit", "cloglog", "cauchit"),
           col = c("limegreen", "purple", "chocolate", "tan"), lwd = mylwd)
  } else
    abline(v = 0.5, lty = "dashed")
}

for (d in 0) {
  matplot(y, cbind(logit(y, deriv = d, inverse = TRUE),
                   probit(y, deriv = d, inverse = TRUE)), las = 1,
          type = "n", col = "purple", xlab = "transformation", ylab = "p",
          main = if (d ==  0) "Some inverse probability link functions"
          else "First derivative")
  lines(y,   logit(y, deriv = d, inverse = TRUE), col = "limegreen")
  lines(y,  probit(y, deriv = d, inverse = TRUE), col = "purple")
  lines(y, cloglog(y, deriv = d, inverse = TRUE), col = "chocolate")
  lines(y, cauchit(y, deriv = d, inverse = TRUE), col = "tan")
  if (d ==  0) {
    abline(h = 0.5, v = 0, lty = "dashed")
    legend(-4, 1, c("logit", "probit", "cloglog", "cauchit"),
           col = c("limegreen", "purple", "chocolate", "tan"), lwd = mylwd)
  }
}

p <- seq(0.21, 0.59, by = 0.01)
plot(p, extlogit(p, min = 0.2, max = 0.6),
     type = "l", col = "black", ylab = "transformation", xlim = c(0, 1),
     las = 1, main = "extlogit(p, min = 0.2, max = 0.6)")
par(lwd = 1)

## End(Not run)

[Package VGAM version 1.0-6 Index]