loglapUC {VGAM}	R Documentation

The Log-Laplace Distribution

Description

Density, distribution function, quantile function and random generation for the 3-parameter log-Laplace distribution with location parameter location.ald, scale parameter scale.ald (on the log scale), and asymmetry parameter kappa.

Usage

dloglap(x, location.ald = 0, scale.ald = 1,
        tau = 0.5, kappa = sqrt(tau/(1-tau)), log = FALSE)
ploglap(q, location.ald = 0, scale.ald = 1,
        tau = 0.5, kappa = sqrt(tau/(1-tau)), lower.tail = TRUE, log.p = FALSE)
qloglap(p, location.ald = 0, scale.ald = 1,
        tau = 0.5, kappa = sqrt(tau/(1-tau)), lower.tail = TRUE, log.p = FALSE)
rloglap(n, location.ald = 0, scale.ald = 1,
        tau = 0.5, kappa = sqrt(tau/(1-tau)))

Arguments

x, q
vector of quantiles.

p
vector of probabilities.

n
number of observations. If length(n) > 1 then the length is taken to be the number required.

location.ald, scale.ald
the location parameter xi and the (positive) scale parameter sigma, on the log scale.

tau
the quantile parameter tau. Must consist of values in (0,1). This argument is used to specify kappa and is ignored if kappa is assigned.

kappa
the asymmetry parameter kappa. Must consist of positive values.

log
if TRUE, probabilities p are given as log(p).

lower.tail, log.p
Same meaning as in pnorm or qnorm.

Details

A positive random variable Y is said to have a log-Laplace distribution if \log(Y) has an asymmetric Laplace distribution (ALD). There are many variants of ALDs and the one used here is described in alaplace3.

Value

dloglap gives the density, ploglap gives the distribution function, qloglap gives the quantile function, and rloglap generates random deviates.

Author(s)

T. W. Yee and Kai Huang

References

Kozubowski, T. J. and Podgorski, K. (2003) Log-Laplace distributions. International Mathematical Journal, 3, 467–495.

See Also

dalap, alaplace3, loglaplace1.

Examples

loc <- 0; sigma <- exp(0.5); kappa <- 1
x <- seq(-0.2, 5, by = 0.01)
## Not run: 
plot(x, dloglap(x, loc, sigma, kappa = kappa), type = "l", col = "blue",
     main = "Blue is density, red is cumulative distribution function",
     ylim = c(0,1), sub = "Purple are 5,10,...,95 percentiles", las = 1, ylab = "")
abline(h = 0, col = "blue", lty = 2)
lines(qloglap(seq(0.05,0.95,by = 0.05), loc, sigma, kappa = kappa),
      dloglap(qloglap(seq(0.05,0.95,by = 0.05), loc, sigma, kappa = kappa),
                loc, sigma, kappa = kappa), col = "purple", lty = 3, type = "h")
lines(x, ploglap(x, loc, sigma, kappa = kappa), type = "l", col = "red")
abline(h = 0, lty = 2)

## End(Not run)
ploglap(qloglap(seq(0.05,0.95,by = 0.05), loc, sigma, kappa = kappa),
          loc, sigma, kappa = kappa)

[Package VGAM version 1.0-6 Index]