loglog {VGAM}	R Documentation

Log-log Link Function

Description

Computes the log-log transformation, including its inverse and the first two derivatives.

Usage

loglog(theta, bvalue = NULL, inverse = FALSE, deriv = 0,
       short = TRUE, tag = FALSE)

Arguments

theta
Numeric or character. See below for further details.

bvalue
Values of theta which are less than or equal to 1 can be replaced by bvalue before computing the link function value. The component name bvalue stands for “boundary value”. See Links for more information.

inverse, deriv, short, tag
Details at Links.

Details

The log-log link function is commonly used for parameters that are greater than unity. Numerical values of theta close to 1 or out of range result in Inf, -Inf, NA or NaN.

Value

For deriv = 0, the log of theta, i.e., log(log(theta)) when inverse = FALSE, and if inverse = TRUE then exp(exp(theta)).

For deriv = 1, then the function returns d theta / d eta as a function of theta if inverse = FALSE, else if inverse = TRUE then it returns the reciprocal.

Here, all logarithms are natural logarithms, i.e., to base e.

Note

Numerical instability may occur when theta is close to 1 unless bvalue is used.

Author(s)

Thomas W. Yee

References

McCullagh, P. and Nelder, J. A. (1989) Generalized Linear Models, 2nd ed. London: Chapman & Hall.

See Also

Links, loge, logoff.

Examples

x <- seq(0.8, 1.5, by = 0.1)
loglog(x)  # Has NAs
loglog(x, bvalue = 1.0 + .Machine$double.eps)  # Has no NAs

x <- seq(1.01, 10, len = 100)
loglog(x)
max(abs(loglog(loglog(x), inverse = TRUE) - x))  # Should be 0

[Package VGAM version 1.0-6 Index]