logoff {VGAM}	R Documentation

Log Link Function with an Offset

Description

Computes the log transformation with an offset, including its inverse and the first two derivatives.

Usage

logoff(theta, offset = 0, inverse = FALSE, deriv = 0,
       short = TRUE, tag = FALSE)

Arguments

theta
Numeric or character. See below for further details.

offset
Offset value. See Links.

inverse, deriv, short, tag
Details at Links.

Details

The log-offset link function is very commonly used for parameters that are greater than a certain value. In particular, it is defined by log(theta + offset) where offset is the offset value. For example, if offset = 0.5 then the value of theta is restricted to be greater than -0.5.

Numerical values of theta close to -offset or out of range result in Inf, -Inf, NA or NaN.

Value

For deriv = 0, the log of theta+offset, i.e., log(theta+offset) when inverse = FALSE, and if inverse = TRUE then exp(theta)-offset.

For deriv = 1, then the function returns d theta / d eta as a function of theta if inverse = FALSE, else if inverse = TRUE then it returns the reciprocal.

Here, all logarithms are natural logarithms, i.e., to base e.

Note

The default means this function is identical to loge.

Numerical instability may occur when theta is close to -offset.

Author(s)

Thomas W. Yee

References

McCullagh, P. and Nelder, J. A. (1989) Generalized Linear Models, 2nd ed. London: Chapman & Hall.

See Also

Links, loge.

Examples

## Not run: 
logoff(seq(-0.2, 0.5, by = 0.1))
logoff(seq(-0.2, 0.5, by = 0.1), offset = 0.5)
   log(seq(-0.2, 0.5, by = 0.1) + 0.5) 
## End(Not run)

[Package VGAM version 1.0-6 Index]