lomax {VGAM}	R Documentation

Lomax Distribution Family Function

Description

Maximum likelihood estimation of the 2-parameter Lomax distribution.

Usage

lomax(lscale = "loge", lshape3.q = "loge", iscale = NULL,
      ishape3.q = NULL, imethod = 1, gscale = exp(-5:5),
      gshape3.q = seq(0.75, 4, by = 0.25),
      probs.y = c(0.25, 0.5, 0.75), zero = "shape")

Arguments

lscale, lshape3.q
Parameter link function applied to the (positive) parameters scale and q. See Links for more choices.

iscale, ishape3.q, imethod
See CommonVGAMffArguments for information. For imethod = 2 a good initial value for iscale is needed to obtain a good estimate for the other parameter.

gscale, gshape3.q, zero, probs.y
See CommonVGAMffArguments.

Details

The 2-parameter Lomax distribution is the 4-parameter generalized beta II distribution with shape parameters a=p=1. It is probably more widely known as the Pareto (II) distribution. It is also the 3-parameter Singh-Maddala distribution with shape parameter a=1, as well as the beta distribution of the second kind with p=1. More details can be found in Kleiber and Kotz (2003).

The Lomax distribution has density

f(y) = q / [b (1 + y/b)^(1+q)]

for b > 0, q > 0, y >= 0. Here, b is the scale parameter scale, and q is a shape parameter. The cumulative distribution function is

F(y) = 1 - [1 + (y/b)]^(-q).

The mean is

E(Y) = b/(q-1)

provided q > 1; these are returned as the fitted values. This family function handles multiple responses.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Note

See the notes in genbetaII.

Author(s)

T. W. Yee

References

Kleiber, C. and Kotz, S. (2003) Statistical Size Distributions in Economics and Actuarial Sciences, Hoboken, NJ, USA: Wiley-Interscience.

See Also

Lomax, genbetaII, betaII, dagum, sinmad, fisk, inv.lomax, paralogistic, inv.paralogistic, simulate.vlm.

Examples

ldata <- data.frame(y = rlomax(n = 1000, scale =  exp(1), exp(2)))
fit <- vglm(y ~ 1, lomax, data = ldata, trace = TRUE)
coef(fit, matrix = TRUE)
Coef(fit)
summary(fit)

[Package VGAM version 1.0-6 Index]