marital.nz {VGAM}	R Documentation

New Zealand Marital Data

Description

Some marital data mainly from a large NZ company collected in the early 1990s.

Usage

data(marital.nz)

Format

A data frame with 6053 observations on the following 3 variables.

age

a numeric vector, age in years

ethnicity

a factor with levels European Maori Other Polynesian. Only Europeans are included in the data set.

mstatus

a factor with levels Divorced/Separated, Married/Partnered, Single, Widowed.

Details

This is a subset of a data set collected from a self-administered questionnaire administered in a large New Zealand workforce observational study conducted during 1992–3. The data were augmented by a second study consisting of retirees. The data can be considered a reasonable representation of the white male New Zealand population in the early 1990s.

Source

Clinical Trials Research Unit, University of Auckland, New Zealand.

References

See bmi.nz and chest.nz.

Examples

summary(marital.nz)

[Package VGAM version 1.0-6 Index]