maxwell {VGAM}	R Documentation

Maxwell Distribution Family Function

Description

Estimating the parameter of the Maxwell distribution by maximum likelihood estimation.

Usage

maxwell(link = "loge", zero = NULL)

Arguments

link, zero
Parameter link function applied to a, which is called the parameter rate. See Links for more choices and information; a log link is the default because the parameter is positive. More information is at CommonVGAMffArguments.

Details

The Maxwell distribution, which is used in the area of thermodynamics, has a probability density function that can be written

f(y;a) = sqrt(2/pi) * a^(3/2) * y^2 * exp(-0.5*a*y^2)

for y>0 and a>0. The mean of Y is sqrt(8 / (a * pi)) (returned as the fitted values), and its variance is (3*pi - 8)/(pi*a).

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, rrvglm and vgam.

Note

Fisher-scoring and Newton-Raphson are the same here. A related distribution is the Rayleigh distribution. This VGAM family function handles multiple responses. This VGAM family function can be mimicked by poisson.points(ostatistic = 1.5, dimension = 2).

Author(s)

T. W. Yee

References

von Seggern, D. H. (1993) CRC Standard Curves and Surfaces, Boca Raton, FL.: CRC Press.

See Also

Maxwell, rayleigh, poisson.points.

Examples

mdata <- data.frame(y = rmaxwell(1000, rate = exp(2)))
fit <- vglm(y ~ 1, maxwell, data = mdata, trace = TRUE, crit = "coef")
coef(fit, matrix = TRUE)
Coef(fit)

[Package VGAM version 1.0-6 Index]