mccullagh89 {VGAM}	R Documentation

McCullagh (1989) Distribution Family Function

Description

Estimates the two parameters of the McCullagh (1989) distribution by maximum likelihood estimation.

Usage

mccullagh89(ltheta = "rhobit", lnu = logoff(offset = 0.5),
            itheta = NULL, inu = NULL, zero = NULL)

Arguments

ltheta, lnu
Link functions for the theta and nu parameters. See Links for general information.

itheta, inu
Numeric. Optional initial values for theta and nu. The default is to internally compute them.

zero
See CommonVGAMffArguments for information.

Details

The McCullagh (1989) distribution has density function

f(y;theta,nu) = (1-y^2)^(nu-0.5) / [ (1 - 2*theta*y+theta^2)^nu * Beta(nu+0.5, 0.5)]

where -1 < y < 1 and -1 < theta < 1. This distribution is equation (1) in that paper. The parameter nu satisfies nu > -1/2, therefore the default is to use an log-offset link with offset equal to 0.5, i.e., eta_2=log(nu+0.5). The mean is of Y is nu*theta/(1+nu), and these are returned as the fitted values.

This distribution is related to the Leipnik distribution (see Johnson et al. (1995)), is related to ultraspherical functions, and under certain conditions, arises as exit distributions for Brownian motion. Fisher scoring is implemented here and it uses a diagonal matrix so the parameters are globally orthogonal in the Fisher information sense. McCullagh (1989) also states that, to some extent, theta and nu have the properties of a location parameter and a precision parameter, respectively.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, rrvglm and vgam.

Note

Convergence may be slow or fail unless the initial values are reasonably close. If a failure occurs, try assigning the argument inu and/or itheta. Figure 1 of McCullagh (1989) gives a broad range of densities for different values of theta and nu, and this could be consulted for obtaining reasonable initial values if all else fails.

Author(s)

T. W. Yee

References

McCullagh, P. (1989) Some statistical properties of a family of continuous univariate distributions. Journal of the American Statistical Association, 84, 125–129.

Johnson, N. L. and Kotz, S. and Balakrishnan, N. (1995) Continuous Univariate Distributions, 2nd edition, Volume 2, New York: Wiley. (pages 612–617).

See Also

leipnik, rhobit, logoff.

Examples

mdata <- data.frame(y = rnorm(n = 1000, sd = 0.2))  # Limit as theta = 0, nu = Inf
fit <- vglm(y ~ 1, mccullagh89, data = mdata, trace = TRUE)
head(fitted(fit))
with(mdata, mean(y))
summary(fit)
coef(fit, matrix = TRUE)
Coef(fit)

[Package VGAM version 1.0-6 Index]