mills.ratio {VGAM}	R Documentation

Mills Ratio

Description

Computes the Mills ratio.

Usage

mills.ratio(x)
mills.ratio2(x)

Arguments

x
Numeric (real).

Details

The Mills ratio here is dnorm(x) / pnorm(x) (some use (1 - pnorm(x)) / dnorm(x)). Some care is needed as x approaches -Inf; when x is very negative then its value approaches -x.

Value

mills.ratio returns the Mills ratio, and mills.ratio2 returns dnorm(x) * dnorm(x) / pnorm(x).

Author(s)

T. W. Yee

References

Mills, J. P. (1926) Table of the ratio: area to bounding ordinate, for any portion of normal curve. Biometrika. 18(3/4), 395–400.

See Also

Normal, tobit, cens.poisson.

Examples

## Not run: 
curve(mills.ratio, -5, 5, col = "orange", las = 1)
curve(mills.ratio, -5, 5, col = "orange", las = 1, log = "y")

## End(Not run)

[Package VGAM version 1.0-6 Index]