notdocumentedyet {VGAM}	R Documentation

Undocumented and Internally Used Functions and Classes

Description

Those currently undocumented and internally used functions are aliased to this help file. Ditto for some classes.

Details

In the VGAM package there are currently many objects/methods/classes which are currently internal and/or undocumented. The help file suppresses the warnings when the package is 'CHECK'ed.

Value

Each objects/methods/classes may or may not have its own individual value. These will be documented over time.

Author(s)

T. W. Yee

[Package VGAM version 1.0-6 Index]