otzeta {VGAM}	R Documentation

One-truncated Zeta Distribution Family Function

Description

Estimates the parameter of the 1-truncated zeta distribution.

Usage

otzeta(lshape = "loge", ishape = NULL, gshape = exp((-4:3)/4), zero = NULL)

Arguments

lshape, ishape, gshape, zero
Same as zetaff.

Details

The 1-truncated zeta distribution is the ordinary zeta distribution but with the probability of one being 0. Thus the other probabilities are scaled up (i.e., divided by 1-P[Y=1]). The mean is returned by default as the fitted values. More details can be found at zetaff. Multiple responses are handled.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Author(s)

T. W. Yee

See Also

Otzeta, zetaff, oizeta, diffzeta, zeta, dzeta, hzeta, zipf.

Examples

odata <- data.frame(x2 = runif(nn <- 1000))  # Artificial data
odata <- transform(odata, shape = loge(-0.25 + x2, inverse = TRUE))
odata <- transform(odata, y1 = rotzeta(nn, shape))
with(odata, table(y1))
ofit <- vglm(y1 ~ x2, otzeta, data = odata, trace = TRUE, crit = "coef")
coef(ofit, matrix = TRUE)

[Package VGAM version 1.0-6 Index]