paralogistic {VGAM}	R Documentation

Paralogistic Distribution Family Function

Description

Maximum likelihood estimation of the 2-parameter paralogistic distribution.

Usage

paralogistic(lscale = "loge", lshape1.a = "loge", iscale = NULL,
    ishape1.a = NULL, imethod = 1, lss = TRUE, gscale = exp(-5:5),
    gshape1.a = seq(0.75, 4, by = 0.25), probs.y = c(0.25, 0.5, 0.75),
    zero = "shape")

Arguments

lss
See CommonVGAMffArguments for important information.

lshape1.a, lscale
Parameter link functions applied to the (positive) parameters a and scale. See Links for more choices.

iscale, ishape1.a, imethod, zero
See CommonVGAMffArguments for information. For imethod = 2 a good initial value for ishape1.a is needed to obtain good estimates for the other parameter.

gscale, gshape1.a
See CommonVGAMffArguments for information.

probs.y
See CommonVGAMffArguments for information.

Details

The 2-parameter paralogistic distribution is the 4-parameter generalized beta II distribution with shape parameter p=1 and a=q. It is the 3-parameter Singh-Maddala distribution with a=q. More details can be found in Kleiber and Kotz (2003).

The 2-parameter paralogistic has density

f(y) = a^2 y^(a-1) / [b^a (1 + (y/b)^a)^(1+a)]

for a > 0, b > 0, y >= 0. Here, b is the scale parameter scale, and a is the shape parameter. The mean is

E(Y) = b gamma(1 + 1/a) gamma(a - 1/a) / gamma(a)

provided a > 1; these are returned as the fitted values. This family function handles multiple responses.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Note

See the notes in genbetaII.

Author(s)

T. W. Yee

References

Kleiber, C. and Kotz, S. (2003) Statistical Size Distributions in Economics and Actuarial Sciences, Hoboken, NJ, USA: Wiley-Interscience.

See Also

Paralogistic, sinmad, genbetaII, betaII, dagum, fisk, inv.lomax, lomax, inv.paralogistic.

Examples

pdata <- data.frame(y = rparalogistic(n = 3000, exp(1), scale = exp(1)))
fit <- vglm(y ~ 1, paralogistic(lss = FALSE), data = pdata, trace = TRUE)
fit <- vglm(y ~ 1, paralogistic(ishape1.a = 2.3, iscale = 5),
            data = pdata, trace = TRUE)
coef(fit, matrix = TRUE)
Coef(fit)
summary(fit)

[Package VGAM version 1.0-6 Index]