paretoIV {VGAM}	R Documentation

Pareto(IV/III/II) Distribution Family Functions

Description

Estimates three of the parameters of the Pareto(IV) distribution by maximum likelihood estimation. Some special cases of this distribution are also handled.

Usage

paretoIV(location = 0, lscale = "loge", linequality = "loge", lshape = "loge",
         iscale = 1, iinequality = 1, ishape = NULL, imethod = 1)
paretoIII(location = 0, lscale = "loge", linequality = "loge",
          iscale = NULL, iinequality = NULL)
paretoII(location = 0, lscale = "loge", lshape = "loge",
         iscale = NULL, ishape = NULL)

Arguments

location
Location parameter, called a below. It is assumed known.

lscale, linequality, lshape
Parameter link functions for the scale parameter (called b below), inequality parameter (called g below), and shape parameter (called s below). See Links for more choices. A log link is the default for all because all these parameters are positive.

iscale, iinequality, ishape
Initial values for the parameters. A NULL value means that it is obtained internally. If convergence failure occurs, use these arguments to input some alternative initial values.

imethod
Method of initialization for the shape parameter. Currently only values 1 and 2 are available. Try the other value if convergence failure occurs.

Details

The Pareto(IV) distribution, which is used in actuarial science, economics, finance and telecommunications, has a cumulative distribution function that can be written

F(y) = 1 - [1 + ((y-a)/b)^(1/g)]^(-s)

for y > a, b>0, g>0 and s>0. The a is called the location parameter, b the scale parameter, g the inequality parameter, and s the shape parameter.

The location parameter is assumed known otherwise the Pareto(IV) distribution will not be a regular family. This assumption is not too restrictive in modelling because in typical applications this parameter is known, e.g., in insurance and reinsurance it is pre-defined by a contract and can be represented as a deductible or a retention level.

The inequality parameter is so-called because of its interpretation in the economics context. If we choose a unit shape parameter value and a zero location parameter value then the inequality parameter is the Gini index of inequality, provided g<=1.

The fitted values are currently the median, e.g., qparetoIV is used for paretoIV().

There are a number of special cases of the Pareto(IV) distribution. These include the Pareto(I), Pareto(II), Pareto(III), and Burr family of distributions. Denoting PIV(a,b,g,s) as the Pareto(IV) distribution, the Burr distribution Burr(b,g,s) is PIV(a=0,b,1/g,s), the Pareto(III) distribution PIII(a,b,g) is PIV(a,b,g,s=1), the Pareto(II) distribution PII(a,b,s) is PIV(a,b,g=1,s), and the Pareto(I) distribution PI(b,s) is PIV(b,b,g=1,s). Thus the Burr distribution can be fitted using the negloge link function and using the default location=0 argument. The Pareto(I) distribution can be fitted using paretoff but there is a slight change in notation: s=k and b=alpha.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Warning

The Pareto(IV) distribution is very general, for example, special cases include the Pareto(I), Pareto(II), Pareto(III), and Burr family of distributions. [Johnson et al. (1994) says on p.19 that fitting Type IV by ML is very difficult and rarely attempted]. Consequently, reasonably good initial values are recommended, and convergence to a local solution may occur. For this reason setting trace=TRUE is a good idea for monitoring the convergence. Large samples are ideally required to get reasonable results.

Note

The extra slot of the fitted object has a component called "location" which stores the location parameter value(s).

Author(s)

T. W. Yee

References

Johnson N. L., Kotz S., and Balakrishnan N. (1994) Continuous Univariate Distributions, Volume 1, 2nd ed. New York: Wiley.

Brazauskas, V. (2003) Information matrix for Pareto(IV), Burr, and related distributions. Comm. Statist. Theory and Methods 32, 315–325.

Arnold, B. C. (1983) Pareto Distributions. Fairland, Maryland: International Cooperative Publishing House.

See Also

ParetoIV, paretoff, gpd.

Examples

pdata <- data.frame(y = rparetoIV(2000, scale = exp(1),
                                  ineq = exp(-0.3), shape = exp(1)))
## Not run: par(mfrow = c(2, 1))
with(pdata, hist(y)); with(pdata, hist(log(y))) 
## End(Not run)
fit <- vglm(y ~ 1, paretoIV, data = pdata, trace = TRUE)
head(fitted(fit))
summary(pdata)
coef(fit, matrix = TRUE)
Coef(fit)
summary(fit)

[Package VGAM version 1.0-6 Index]