pgamma.deriv.unscaled {VGAM}	R Documentation

Derivatives of the Incomplete Gamma Integral (Unscaled Version)

Description

The first two derivatives of the incomplete gamma integral with scaling.

Usage

pgamma.deriv.unscaled(q, shape)

Arguments

q, shape
As in pgamma and pgamma.deriv but these must be vectors of positive values only and finite.

Details

Define

G(a,x) = integral_0^x t^(a-1) exp(-t) dt

so that G(x, a) is pgamma(x, a) * gamma(a). Write x = q and shape = a. The 0th and first and second derivatives with respect to a of G are returned. This function is similar in spirit to pgamma.deriv but here there is no gamma function to scale things. Currently a 3-column matrix is returned (in the future this may change and an argument may be supplied so that only what is required by the user is computed.) This function is based on Wingo (1989).

Value

The 3 columns, running from left to right, are the 0:2th derivatives with respect to a.

Warning

These function seems inaccurate for q = 1 and q = 2; see the plot below.

Author(s)

T. W. Yee.

References

See truncweibull.

See Also

pgamma.deriv, pgamma.

Examples

x <- 3; aa <- seq(0.3, 04, by = 0.01)
ans.u <- pgamma.deriv.unscaled(x, aa)
head(ans.u)

## Not run:  par(mfrow = c(1, 3))
for (jay in 1:3) {
  plot(aa, ans.u[, jay], type = "l", col = "blue", cex.lab = 1.5,
       cex.axis = 1.5, las = 1, main = colnames(ans.u)[jay],
       log = "", xlab = "shape", ylab = "")
  abline(h = 0, v = 1:2, lty = "dashed", col = "gray")  # Inaccurate at 1 and 2
}

## End(Not run)

[Package VGAM version 1.0-6 Index]