plotvgam.control {VGAM}	R Documentation

Control Function for plotvgam()

Description

Provides default values for many arguments available for plotvgam().

Usage

plotvgam.control(which.cf = NULL,
                 xlim = NULL, ylim = NULL, llty = par()$lty,
                 slty = "dashed", pcex = par()$cex,
                 pch = par()$pch, pcol = par()$col,
                 lcol = par()$col, rcol = par()$col,
                 scol = par()$col, llwd = par()$lwd, slwd = par()$lwd,
                 add.arg = FALSE, one.at.a.time = FALSE,
                 .include.dots = TRUE, noxmean = FALSE,
                 shade = FALSE, shcol = "gray80", ...)

Arguments

which.cf
Integer vector specifying which component functions are to be plotted (for each covariate). Must have values from the set {1,2,...,M}.

xlim
Range for the x-axis.

ylim
Range for the y-axis.

llty
Line type for the fitted functions (lines). Fed into par(lty).

slty
Line type for the standard error bands. Fed into par(lty).

pcex
Character expansion for the points (residuals). Fed into par(cex).

pch
Character used for the points (residuals). Same as par(pch).

pcol
Color of the points. Fed into par(col).

lcol
Color of the fitted functions (lines). Fed into par(col).

rcol
Color of the rug plot. Fed into par(col).

scol
Color of the standard error bands. Fed into par(col).

llwd
Line width of the fitted functions (lines). Fed into par(lwd).

slwd
Line width of the standard error bands. Fed into par(lwd).

add.arg
Logical. If TRUE then the plot will be added to an existing plot, otherwise a new plot will be made.

one.at.a.time
Logical. If TRUE then the plots are done one at a time, with the user having to hit the return key between the plots.

.include.dots
Not to be used by the user.

noxmean
Logical. If TRUE then the point at the mean of x, which is added when standard errors are specified and it thinks the function is linear, is not added. One might use this argument if ylab is specified.

shade, shcol
shade is logical; if TRUE then the pointwise SE band is shaded gray by default. The colour can be adjusted by setting shcol. These arguments are ignored unless se = TRUE and overlay = FALSE; If shade = TRUE then scol is ignored.

...
Other arguments that may be fed into par().

In the above, M is the number of linear/additive predictors.

Details

The most obvious features of plotvgam can be controlled by the above arguments.

Value

A list with values matching the arguments.

Author(s)

Thomas W. Yee

References

Yee, T. W. and Wild, C. J. (1996) Vector generalized additive models. Journal of the Royal Statistical Society, Series B, Methodological, 58, 481–493.

See Also

plotvgam.

Examples

plotvgam.control(lcol = c("red", "blue"), scol = "darkgreen", se = TRUE)

[Package VGAM version 1.0-6 Index]