pneumo {VGAM}	R Documentation

Pneumoconiosis in Coalminers Data

Description

The pneumo data frame has 8 rows and 4 columns. Exposure time is explanatory, and there are 3 ordinal response variables.

Usage

data(pneumo)

Format

This data frame contains the following columns:

exposure.time

a numeric vector, in years

normal

a numeric vector, counts

mild

a numeric vector, counts

severe

a numeric vector, counts

Details

These were collected from coalface workers. In the original data set, the two most severe categories were combined.

Source

Ashford, J.R., 1959. An approach to the analysis of data for semi-quantal responses in biological assay. Biometrics, 15, 573–581.

References

McCullagh, P. and Nelder, J. A. (1989) Generalized Linear Models, 2nd ed. London: Chapman & Hall.

See Also

cumulative.

Examples

# Fit the proportional odds model, p.179, in McCullagh and Nelder (1989)
pneumo <- transform(pneumo, let = log(exposure.time))
vglm(cbind(normal, mild, severe) ~ let, propodds, data = pneumo)

[Package VGAM version 1.0-6 Index]