PoissonPoints {VGAM}	R Documentation

Poisson Points Distribution

Description

Density for the PoissonPoints distribution.

Usage

dpois.points(x, lambda, ostatistic, dimension = 2, log = FALSE)

Arguments

x
vector of quantiles.

lambda
the mean density of points.

ostatistic
positive values, usually integers.

dimension
Either 2 and/or 3.

log
Logical; if TRUE, the logarithm is returned.

Details

See poisson.points, the VGAM family function for estimating the parameters, for the formula of the probability density function and other details.

Value

dpois.points gives the density.

See Also

poisson.points, dpois, Maxwell.

Examples

## Not run:  lambda <- 1; xvec <- seq(0, 2, length = 400)
plot(xvec, dpois.points(xvec, lambda, ostat = 1, dimension = 2),
     type = "l", las = 1, col = "blue",
     sub = "First order statistic",
     main = paste("PDF of PoissonPoints distribution with lambda = ",
                  lambda, " and on the plane", sep = "")) 
## End(Not run)

[Package VGAM version 1.0-6 Index]