prinia {VGAM}	R Documentation

Yellow-bellied Prinia

Description

A data frame with yellow-bellied Prinia.

Usage

data(prinia)

Format

A data frame with 151 observations on the following 23 variables.

length

a numeric vector, the scaled wing length (zero mean and unit variance).

fat

a numeric vector, fat index; originally 1 (no fat) to 4 (very fat) but converted to 0 (no fat) versus 1 otherwise.

cap

a numeric vector, number of times the bird was captured or recaptured.

noncap

a numeric vector, number of times the bird was not captured.

y01, y02, y03, y04, y05, y06

a numeric vector of 0s and 1s; for noncapture and capture resp.

y07, y08, y09, y10, y11, y12

same as above.

y13, y14, y15, y16, y17, y18, y19

same as above.

Details

The yellow-bellied Prinia Prinia flaviventris is a common bird species located in Southeast Asia. A capture–recapture experiment was conducted at the Mai Po Nature Reserve in Hong Kong during 1991, where captured individuals had their wing lengths measured and fat index recorded. A total of 19 weekly capture occasions were considered, where 151 distinct birds were captured.

More generally, the prinias are a genus of small insectivorous birds, and are sometimes referred to as wren-warblers. They are a little-known group of the tropical and subtropical Old World, the roughly 30 species being divided fairly equally between Africa and Asia.

Source

Thanks to Paul Yip for permission to make this data available.

Hwang, W.-H. and Huggins, R. M. (2007) Application of semiparametric regression models in the analysis of capture–recapture experiments. Australian and New Zealand Journal of Statistics 49, 191–202.

Examples

head(prinia)
summary(prinia)
rowSums(prinia[, c("cap", "noncap")])  # 19s

#  Fit a positive-binomial distribution (M.h) to the data:
fit1 <- vglm(cbind(cap, noncap) ~ length + fat, posbinomial, data = prinia)

#  Fit another positive-binomial distribution (M.h) to the data:
#  The response input is suitable for posbernoulli.*-type functions.
fit2 <- vglm(cbind(y01, y02, y03, y04, y05, y06, y07, y08, y09, y10,
                   y11, y12, y13, y14, y15, y16, y17, y18, y19) ~
             length + fat, posbernoulli.b(drop.b = FALSE ~ 0), data = prinia)

[Package VGAM version 1.0-6 Index]