put.smart {VGAM}	R Documentation

Adds a List to the End of the List “.smart.prediction”

Description

Adds a list to the end of the list .smart.prediction in smartpredenv.

Usage

put.smart(smart)

Arguments

smart
a list containing parameters needed later for smart prediction.

Details

put.smart is used in "write" mode within a smart function. It saves parameters at the time of model fitting, which are later used for prediction. The function put.smart is the opposite of get.smart, and both deal with the same contents.

Value

Nothing is returned.

Side Effects

The variable .smart.prediction.counter in smartpredenv is incremented beforehand, and .smart.prediction[[.smart.prediction.counter]] is assigned the list smart. If the list .smart.prediction in smartpredenv is not long enough to hold smart, then it is made larger, and the variable .max.smart in smartpredenv is adjusted accordingly.

See Also

get.smart.

Examples

print(sm.min1)

[Package VGAM version 1.0-6 Index]