rec.exp1 {VGAM}	R Documentation

Upper Record Values from a 1-parameter Exponential Distribution

Description

Maximum likelihood estimation of the rate parameter of a 1-parameter exponential distribution when the observations are upper record values.

Usage

rec.exp1(lrate = "loge", irate = NULL, imethod = 1)

Arguments

lrate
Link function applied to the rate parameter. See Links for more choices.

irate
Numeric. Optional initial values for the rate. The default value NULL means they are computed internally, with the help of imethod.

imethod
Integer, either 1 or 2 or 3. Initial method, three algorithms are implemented. Choose the another value if convergence fails, or use irate.

Details

The response must be a vector or one-column matrix with strictly increasing values.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Note

By default, this family function has the intercept-only MLE as the initial value, therefore convergence may only take one iteration. Fisher scoring is used.

Author(s)

T. W. Yee

References

Arnold, B. C. and Balakrishnan, N. and Nagaraja, H. N. (1998) Records, New York: John Wiley & Sons.

See Also

exponential.

Examples

rawy <- rexp(n <- 10000, rate = exp(1))
y <- unique(cummax(rawy))  # Keep only the records

length(y) / y[length(y)]   # MLE of rate

fit <- vglm(y ~ 1, rec.exp1, trace = TRUE)
coef(fit, matrix = TRUE)
Coef(fit)

[Package VGAM version 1.0-6 Index]