reciprocal {VGAM}	R Documentation

Reciprocal Link Function

Description

Computes the reciprocal transformation, including its inverse and the first two derivatives.

Usage

   reciprocal(theta, bvalue = NULL, inverse = FALSE, deriv = 0,
              short = TRUE, tag = FALSE)
negreciprocal(theta, bvalue = NULL, inverse = FALSE, deriv = 0,
              short = TRUE, tag = FALSE)

Arguments

theta
Numeric or character. See below for further details.

bvalue
See Links.

inverse, deriv, short, tag
Details at Links.

Details

The reciprocal link function is a special case of the power link function. Numerical values of theta close to 0 result in Inf, -Inf, NA or NaN.

The negreciprocal link function computes the negative reciprocal, i.e., -1/theta.

Value

For reciprocal: for deriv = 0, the reciprocal of theta, i.e., 1/theta when inverse = FALSE, and if inverse = TRUE then 1/theta. For deriv = 1, then the function returns d theta / d eta as a function of theta if inverse = FALSE, else if inverse = TRUE then it returns the reciprocal.

Note

Numerical instability may occur when theta is close to 0.

Author(s)

Thomas W. Yee

References

McCullagh, P. and Nelder, J. A. (1989) Generalized Linear Models, 2nd ed. London: Chapman & Hall.

See Also

identity, powerlink.

Examples

   reciprocal(1:5)
   reciprocal(1:5, inverse = TRUE, deriv = 2)
negreciprocal(1:5)
negreciprocal(1:5, inverse = TRUE, deriv = 2)

x <- (-3):3
reciprocal(x)  # Has Inf
reciprocal(x, bvalue = .Machine$double.eps)  # Has no Inf

[Package VGAM version 1.0-6 Index]