rhobit {VGAM}	R Documentation

Rhobit Link Function

Description

Computes the rhobit link transformation, including its inverse and the first two derivatives.

Usage

rhobit(theta, bminvalue = NULL, bmaxvalue = NULL,
       inverse = FALSE, deriv = 0, short = TRUE, tag = FALSE)

Arguments

theta
Numeric or character. See below for further details.

bminvalue, bmaxvalue
Optional boundary values, e.g., values of theta which are less than or equal to -1 can be replaced by bminvalue before computing the link function value. And values of theta which are greater than or equal to 1 can be replaced by bmaxvalue before computing the link function value. See Links.

inverse, deriv, short, tag
Details at Links.

Details

The rhobit link function is commonly used for parameters that lie between -1 and 1. Numerical values of theta close to -1 or 1 or out of range result in Inf, -Inf, NA or NaN.

Value

For deriv = 0, the rhobit of theta, i.e., log((1 + theta)/(1 - theta)) when inverse = FALSE, and if inverse = TRUE then (exp(theta) - 1)/(exp(theta) + 1).

For deriv = 1, then the function returns d eta / d theta as a function of theta if inverse = FALSE, else if inverse = TRUE then it returns the reciprocal.

Note

Numerical instability may occur when theta is close to -1 or 1. One way of overcoming this is to use bminvalue, etc.

The correlation parameter of a standard bivariate normal distribution lies between -1 and 1, therefore this function can be used for modelling this parameter as a function of explanatory variables.

The link function rhobit is very similar to fisherz, e.g., just twice the value of fisherz.

Author(s)

Thomas W. Yee

See Also

Links, binom2.rho, fisherz.

Examples

theta <- seq(-0.99, 0.99, by = 0.01)
y <- rhobit(theta)
## Not run: 
plot(theta, y, type = "l", las = 1, ylab = "", main = "rhobit(theta)")
abline(v = 0, h = 0, lty = 2)

## End(Not run)

x <- c(seq(-1.02, -0.98, by = 0.01), seq(0.97, 1.02, by = 0.01))
rhobit(x)  # Has NAs
rhobit(x, bminvalue = -1 + .Machine$double.eps,
          bmaxvalue =  1 - .Machine$double.eps)  # Has no NAs

[Package VGAM version 1.0-6 Index]