rigff {VGAM}	R Documentation

Reciprocal Inverse Gaussian distribution

Description

Estimation of the parameters of a reciprocal inverse Gaussian distribution.

Usage

rigff(lmu = "identitylink", llambda = "loge", imu = NULL, ilambda = 1)

Arguments

lmu, llambda
Link functions for mu and lambda. See Links for more choices.

imu, ilambda
Initial values for mu and lambda. A NULL means a value is computed internally.

Details

See Jorgensen (1997) for details.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Note

This distribution is potentially useful for dispersion modelling.

Author(s)

T. W. Yee

References

Jorgensen, B. (1997) The Theory of Dispersion Models. London: Chapman & Hall

See Also

simplex.

Examples

rdata <- data.frame(y = rchisq(n = 100, df = 14))  # Not 'proper' data!!
fit <- vglm(y ~ 1, rigff, data = rdata, trace = TRUE)
fit <- vglm(y ~ 1, rigff, data = rdata, trace = TRUE, eps = 1e-9, crit = "coef")
summary(fit)

[Package VGAM version 1.0-6 Index]