rrvglm-class {VGAM}	R Documentation

Class “rrvglm”

Description

Reduced-rank vector generalized linear models.

Objects from the Class

Objects can be created by calls to rrvglm.

Slots

extra:

Object of class "list"; the extra argument on entry to vglm. This contains any extra information that might be needed by the family function.

family:

Object of class "vglmff". The family function.

iter:

Object of class "numeric". The number of IRLS iterations used.

predictors:

Object of class "matrix" with M columns which holds the M linear predictors.

assign:

Object of class "list", from class "vlm". This named list gives information matching the columns and the (LM) model matrix terms.

call:

Object of class "call", from class "vlm". The matched call.

coefficients:

Object of class "numeric", from class "vlm". A named vector of coefficients.

constraints:

Object of class "list", from class "vlm". A named list of constraint matrices used in the fitting.

contrasts:

Object of class "list", from class "vlm". The contrasts used (if any).

control:

Object of class "list", from class "vlm". A list of parameters for controlling the fitting process. See vglm.control for details.

criterion:

Object of class "list", from class "vlm". List of convergence criterion evaluated at the final IRLS iteration.

df.residual:

Object of class "numeric", from class "vlm". The residual degrees of freedom.

df.total:

Object of class "numeric", from class "vlm". The total degrees of freedom.

dispersion:

Object of class "numeric", from class "vlm". The scaling parameter.

effects:

Object of class "numeric", from class "vlm". The effects.

fitted.values:

Object of class "matrix", from class "vlm". The fitted values. This is usually the mean but may be quantiles, or the location parameter, e.g., in the Cauchy model.

misc:

Object of class "list", from class "vlm". A named list to hold miscellaneous parameters.

model:

Object of class "data.frame", from class "vlm". The model frame.

na.action:

Object of class "list", from class "vlm". A list holding information about missing values.

offset:

Object of class "matrix", from class "vlm". If non-zero, a M-column matrix of offsets.

post:

Object of class "list", from class "vlm" where post-analysis results may be put.

preplot:

Object of class "list", from class "vlm" used by plotvgam; the plotting parameters may be put here.

prior.weights:

Object of class "matrix", from class "vlm" holding the initially supplied weights.

qr:

Object of class "list", from class "vlm". QR decomposition at the final iteration.

R:

Object of class "matrix", from class "vlm". The R matrix in the QR decomposition used in the fitting.

rank:

Object of class "integer", from class "vlm". Numerical rank of the fitted model.

residuals:

Object of class "matrix", from class "vlm". The working residuals at the final IRLS iteration.

ResSS:

Object of class "numeric", from class "vlm". Residual sum of squares at the final IRLS iteration with the adjusted dependent vectors and weight matrices.

smart.prediction:

Object of class "list", from class "vlm". A list of data-dependent parameters (if any) that are used by smart prediction.

terms:

Object of class "list", from class "vlm". The terms object used.

weights:

Object of class "matrix", from class "vlm". The weight matrices at the final IRLS iteration. This is in matrix-band form.

x:

Object of class "matrix", from class "vlm". The model matrix (LM, not VGLM).

xlevels:

Object of class "list", from class "vlm". The levels of the factors, if any, used in fitting.

y:

Object of class "matrix", from class "vlm". The response, in matrix form.

Xm2:

Object of class "matrix", from class "vlm". See vglm-class).

Ym2:

Object of class "matrix", from class "vlm". See vglm-class).

callXm2:

Object of class "call", from class "vlm". The matched call for argument form2.

Extends

Class "vglm", directly. Class "vlm", by class "vglm".

Methods

biplot

signature(x = "rrvglm"): biplot.

Coef

signature(object = "rrvglm"): more detailed coefficients giving A, \bold{B}1, C, etc.

biplot

signature(object = "rrvglm"): biplot.

print

signature(x = "rrvglm"): short summary of the object.

summary

signature(object = "rrvglm"): a more detailed summary of the object.

Note

The slots of "rrvglm" objects are currently identical to "vglm" objects.

Author(s)

Thomas W. Yee

References

Yee, T. W. and Hastie, T. J. (2003) Reduced-rank vector generalized linear models. Statistical Modelling, 3, 15–41.

Yee, T. W. and Wild, C. J. (1996) Vector generalized additive models. Journal of the Royal Statistical Society, Series B, Methodological, 58, 481–493.

See Also

rrvglm, lvplot.rrvglm, vglmff-class.

Examples

## Not run:  # Rank-1 stereotype model of Anderson (1984)
pneumo <- transform(pneumo, let = log(exposure.time),
                            x3  = runif(nrow(pneumo)))  # x3 is unrelated
fit <- rrvglm(cbind(normal, mild, severe) ~ let + x3,
              multinomial, data = pneumo, Rank = 1)
Coef(fit)

## End(Not run)

[Package VGAM version 1.0-6 Index]