ruge {VGAM}	R Documentation

Rutherford-Geiger Polonium Data

Description

Decay counts of polonium recorded by Rutherford and Geiger (1910).

Usage

data(ruge)

Format

This data frame contains the following columns:

counts

a numeric vector, counts or frequencies

number

a numeric vector, the number of decays

Details

These are the radioactive decay counts of polonium recorded by Rutherford and Geiger (1910) representing the number of scintillations in 2608 1/8 minute intervals. For example, there were 57 frequencies of zero counts. The counts can be thought of as being approximately Poisson distributed.

Source

Rutherford, E. and Geiger, H. (1910) The Probability Variations in the Distribution of alpha Particles, Philosophical Magazine, 20, 698–704.

Examples

lambdahat <- with(ruge, weighted.mean(number, w = counts))
(N <- with(ruge, sum(counts)))
with(ruge, cbind(number, counts,
                 fitted = round(N * dpois(number, lam = lambdahat))))

[Package VGAM version 1.0-6 Index]