score.stat {VGAM}	R Documentation

Rao's Score Test Statistics Evaluated at the Null Values

Description

Generic function that computes Rao's score test statistics evaluated at the null values (consequently they do not suffer from the Hauck-Donner effect).

Usage

score.stat(object, ...)
score.stat.vlm(object, values0 = 0, subset = NULL, omit1s = TRUE,
          all.out = FALSE, iterate = TRUE, trace = FALSE, ...)

Arguments

object, values0, subset
Same as in wald.stat.vlm.

omit1s, all.out, iterate
Same as in wald.stat.vlm.

trace
Same as in wald.stat.vlm.

...
Ignored for now.

Details

The (Rao) score test (also known as the Lagrange multiplier test in econometrics) is a third general method for hypothesis testing under a likelihood-based framework (the others are the likelihood ratio test and Wald test; see lrt.stat and wald.stat). Asymptotically, the three tests are equivalent. The Wald test is not invariant to parameterization, and the usual Wald test statistics computed at the estimates make it vulnerable to the Hauck-Donner effect (HDE; see hdeff). This function is similar to wald.stat in that one coefficient is set to 0 (by default) and the other coefficients are iterated by IRLS to get their MLE subject to this constraint.

Value

By default the signed square root of the Rao score statistics are returned. If all.out = TRUE then a list is returned with the following components: score.stat the score statistic, SE0 the standard error of that coefficient, values0 the null values. Approximately, the default score statistics output are standard normal random variates if each null hypothesis is true.

Warning

See wald.stat.vlm.

Author(s)

Thomas W. Yee

See Also

wald.stat, lrt.stat, summaryvglm, summary.glm, vglm, hdeff.

Examples

set.seed(1)
pneumo <- transform(pneumo, let = log(exposure.time),
                            x3 = rnorm(nrow(pneumo)))
(pfit <- vglm(cbind(normal, mild, severe) ~ let + x3,
             propodds, data = pneumo))
score.stat(pfit)  # No HDE here; should be similar to the next line:
coef(summary(pfit))[, "z value"]  # Wald statistics computed at the MLE
summary(pfit, score0 = TRUE)

[Package VGAM version 1.0-6 Index]