Skellam {VGAM}	R Documentation

The Skellam Distribution

Description

Density and random generation for the Skellam distribution.

Usage

dskellam(x, mu1, mu2, log = FALSE)
rskellam(n, mu1, mu2)

Arguments

x
vector of quantiles.

n
number of observations. Same as runif.

mu1, mu2
See skellam

.

log
Logical; if TRUE, the logarithm is returned.

Details

See skellam, the VGAM family function for estimating the parameters, for the formula of the probability density function and other details.

Value

dskellam gives the density, and rskellam generates random deviates.

Warning

Numerical problems may occur for data if mu1 and/or mu2 are large. The normal approximation for this case has not been implemented yet.

See Also

skellam, dpois.

Examples

## Not run:  mu1 <- 1; mu2 <- 2; x <- (-7):7
plot(x, dskellam(x, mu1, mu2), type = "h", las = 1, col = "blue",
     main = paste("Density of Skellam distribution with mu1 = ", mu1,
                  " and mu2 = ", mu2, sep = "")) 
## End(Not run)

[Package VGAM version 1.0-6 Index]