skewnorm {VGAM}	R Documentation

Skew-Normal Distribution

Description

Density and random generation for the univariate skew-normal distribution.

Usage

dskewnorm(x, location = 0, scale = 1, shape = 0, log = FALSE)
rskewnorm(n, location = 0, scale = 1, shape = 0)

Arguments

x
vector of quantiles.

n
number of observations. Same as runif.

location
The location parameter xi. A vector.

scale
The scale parameter w. A positive vector.

shape
The shape parameter. It is called alpha in skewnormal.

log
Logical. If log=TRUE then the logarithm of the density is returned.

Details

See skewnormal, which currently only estimates the shape parameter. More generally here, Z = xi + w * Y where Y has a standard skew-normal distribution (see skewnormal), xi is the location parameter and w is the scale parameter.

Value

dskewnorm gives the density, rskewnorm generates random deviates.

Note

The default values of all three parameters corresponds to the skew-normal being the standard normal distribution.

Author(s)

T. W. Yee

References

http://tango.stat.unipd.it/SN.

See Also

skewnormal.

Examples

## Not run:  N <- 200  # Grid resolution
shape <- 7; x <- seq(-4, 4, len = N)
plot(x, dskewnorm(x, shape = shape), type = "l", col = "blue", las = 1,
     ylab = "", lty = 1, lwd = 2)
abline(v = 0, h = 0, col = "grey")
lines(x, dnorm(x), col = "orange", lty = 2, lwd = 2)
legend("topleft", leg = c(paste("Blue = dskewnorm(x, ", shape,")", sep = ""),
       "Orange = standard normal density"), lty = 1:2, lwd = 2,
       col = c("blue", "orange")) 
## End(Not run)

[Package VGAM version 1.0-6 Index]