slash {VGAM}	R Documentation

Slash Distribution Family Function

Description

Estimates the two parameters of the slash distribution by maximum likelihood estimation.

Usage

slash(lmu = "identitylink", lsigma = "loge",
      imu = NULL, isigma = NULL, gprobs.y = ppoints(8), nsimEIM = 250,
      zero = NULL, smallno = .Machine$double.eps*1000)

Arguments

lmu, lsigma
Parameter link functions applied to the mu and sigma parameters, respectively. See Links for more choices.

imu, isigma
Initial values. A NULL means an initial value is chosen internally. See CommonVGAMffArguments for more information.

gprobs.y
Used to compute the initial values for mu. This argument is fed into the probs argument of quantile to construct a grid, which is used to evaluate the log-likelihood. This must have values between 0 and 1.

nsimEIM, zero
See CommonVGAMffArguments for information.

smallno
Small positive number, used to test for the singularity.

Details

The standard slash distribution is the distribution of the ratio of a standard normal variable to an independent standard uniform(0,1) variable. It is mainly of use in simulation studies. One of its properties is that it has heavy tails, similar to those of the Cauchy.

The general slash distribution can be obtained by replacing the univariate normal variable by a general normal N(mu,sigma) random variable. It has a density that can be written as

f(y) = 1/(2*sigma*sqrt(2*pi)) if y=mu = 1-exp(-(((x-mu)/sigma)^2)/2))/(sqrt(2*pi)*sigma*((x-mu)/sigma)^2) if y!=mu

where mu and sigma are the mean and standard deviation of the univariate normal distribution respectively.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Note

Fisher scoring using simulation is used. Convergence is often quite slow. Numerical problems may occur.

Author(s)

T. W. Yee and C. S. Chee

References

Johnson, N. L. and Kotz, S. and Balakrishnan, N. (1994) Continuous Univariate Distributions, 2nd edition, Volume 1, New York: Wiley.

Kafadar, K. (1982) A Biweight Approach to the One-Sample Problem Journal of the American Statistical Association, 77, 416–424.

See Also

rslash, simulate.vlm.

Examples

## Not run: 
sdata <- data.frame(y = rslash(n = 1000, mu = 4, sigma = exp(2)))
fit <- vglm(y ~ 1, slash, data = sdata, trace = TRUE)
coef(fit, matrix = TRUE)
Coef(fit)
summary(fit)

## End(Not run)

[Package VGAM version 1.0-6 Index]