smart.mode.is {VGAM}	R Documentation

Determine What Mode the Smart Prediction is In

Description

Determine which of three modes the smart prediction is currently in.

Usage

smart.mode.is(mode.arg = NULL)

Arguments

mode.arg
a character string, either "read", "write" or "neutral".

Details

Smart functions such as bs and poly need to know what mode smart prediction is in. If it is in "write" mode then the parameters are saved to .smart.prediction using put.smart. If in "read" mode then the parameters are read in using get.smart. If in "neutral" mode then the smart function behaves like an ordinary function.

Value

If mode.arg is given, then either TRUE or FALSE is returned. If mode.arg is not given, then the mode ("neutral", "read" or "write") is returned. Usually, the mode is "neutral".

See Also

put.smart, bs, poly.

Examples

print(sm.min1)
smart.mode.is()  # Returns "neutral"
smart.mode.is(smart.mode.is())  # Returns TRUE

[Package VGAM version 1.0-6 Index]