summarypvgam {VGAM}	R Documentation

Summarizing Penalized Vector Generalized Additive Model Fits

Description

These functions are all methods for class "pvgam" or summary.pvgam objects.

Usage

summarypvgam(object, dispersion = NULL, digits = options()$digits - 2,
            presid = TRUE)
## S3 method for class 'summary.pvgam'
show(x, quote = TRUE, prefix = "", digits = options()$digits -
    2, signif.stars = getOption("show.signif.stars"))

Arguments

object
an object of class "pvgam", which is the result of a call to vgam with at least one sm.os or sm.ps term.

x
an object of class "summary.pvgam", which is the result of a call to summarypvgam().

dispersion, digits, presid
See summaryvglm.

quote, prefix, signif.stars
See summaryvglm.

Details

This methods function reports a summary more similar to summary.gam from mgcv than summary.gam() from gam. It applies to G2-VGAMs using sm.os and O-splines, else sm.ps and P-splines. In particular, the hypothesis test for whether each sm.os or sm.ps term can be deleted follows quite closely to summary.gam. The p-values from this type of test tend to be biased downwards (too small) and corresponds to p.type = 5. It is hoped in the short future that improved p-values be implemented, somewhat like the default of summary.gam. This methods function was adapted from summary.gam.

Value

summarypvgam returns an object of class "summary.pvgam"; see summary.pvgam-class.

Warning

See sm.os.

See Also

vgam, summaryvgam, summary.pvgam-class, sm.os, sm.ps, summary.glm, summary.lm, summary.gam from mgcv, summaryvgam for G1-VGAMs.

Examples

hfit2 <- vgam(agaaus ~ sm.os(altitude), binomialff, data = hunua)
coef(hfit2, matrix = TRUE)
summary(hfit2)

[Package VGAM version 1.0-6 Index]