summaryvgam {VGAM}	R Documentation

Summarizing Vector Generalized Additive Model Fits

Description

These functions are all methods for class vgam or summary.vgam objects.

Usage

summaryvgam(object, dispersion = NULL, digits = options()$digits - 2,
            presid = TRUE, nopredictors = FALSE)
## S3 method for class 'summary.vgam'
show(x, quote = TRUE, prefix = "",
                            digits = options()$digits-2, nopredictors = NULL)

Arguments

object
an object of class "vgam", which is the result of a call to vgam with at least one s term.

x
an object of class "summary.vgam", which is the result of a call to summaryvgam().

dispersion, digits, presid
See summaryvglm.

quote, prefix, nopredictors
See summaryvglm.

Details

This methods function reports a summary more similar to summary.gam() from gam than summary.gam from mgcv. It applies to G1-VGAMs using s and vector backfitting. In particular, an approximate score test for linearity is conducted for each s term—see Section 4.3.4 of Yee (2015) for details. The p-values from this type of test tend to be biased upwards (too large).

Value

summaryvgam returns an object of class "summary.vgam"; see summary.vgam-class.

See Also

vgam, summary.glm, summary.lm, summary.gam from mgcv, summarypvgam for P-VGAMs.

Examples

hfit <- vgam(agaaus ~ s(altitude, df = 2), binomialff, data = hunua)
summary(hfit)
summary(hfit)@anova  # Table for (approximate) testing of linearity

[Package VGAM version 1.0-6 Index]