tikuv {VGAM}	R Documentation

Short-tailed Symmetric Distribution Family Function

Description

Fits the short-tailed symmetric distribution of Tiku and Vaughan (1999).

Usage

tikuv(d, lmean = "identitylink", lsigma = "loge", isigma = NULL,
      zero = "sigma")

Arguments

d
The d parameter. It must be a single numeric value less than 2. Then h = 2-d>0 is another parameter.

lmean, lsigma
Link functions for the mean and standard deviation parameters of the usual univariate normal distribution (see Details below). They are mu and sigma respectively. See Links for more choices.

isigma
Optional initial value for sigma. A NULL means a value is computed internally.

zero
A vector specifying which linear/additive predictors are modelled as intercept-only. The values can be from the set {1,2}, corresponding respectively to mu, sigma. If zero = NULL then all linear/additive predictors are modelled as a linear combination of the explanatory variables. For many data sets having zero = 2 is a good idea. See CommonVGAMffArguments for information.

Details

The short-tailed symmetric distribution of Tiku and Vaughan (1999) has a probability density function that can be written

f(y) = (K/(sqrt(2*pi)*sigma)) * [1 + (1/(2*h)) * ((y-mu)/sigma)^2]^2 * exp( -0.5 * (y-mu)^2/ sigma^2)

where h=2-d>0, K is a function of h, -Inf < y < Inf, sigma > 0. The mean of Y is E(Y) = mu and this is returned as the fitted values.

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Warning

Under- or over-flow may occur if the data is ill-conditioned, e.g., when d is very close to 2 or approaches -Inf.

Note

The density function is the product of a univariate normal density and a polynomial in the response y. The distribution is bimodal if d>0, else is unimodal. A normal distribution arises as the limit as d approaches -Inf, i.e., as h approaches Inf. Fisher scoring is implemented. After fitting the value of d is stored in @misc with component name d.

Author(s)

Thomas W. Yee

References

Akkaya, A. D. and Tiku, M. L. (2008) Short-tailed distributions and inliers. Test, 17, 282–296.

Tiku, M. L. and Vaughan, D. C. (1999) A family of short-tailed symmetric distributions. Technical report, McMaster University, Canada.

See Also

dtikuv, uninormal.

Examples

m <- 1.0; sigma <- exp(0.5)
tdata <- data.frame(y = rtikuv(n = 1000, d = 1, m = m, s = sigma))
tdata <- transform(tdata, sy = sort(y))
fit <- vglm(y ~ 1, tikuv(d = 1), data = tdata, trace = TRUE)
coef(fit, matrix = TRUE)
(Cfit <- Coef(fit))
with(tdata, mean(y))
## Not run:  with(tdata, hist(y, prob = TRUE))
lines(dtikuv(sy, d = 1, m = Cfit[1], s = Cfit[2]) ~ sy, data = tdata, col = "orange") 
## End(Not run)

[Package VGAM version 1.0-6 Index]