topple {VGAM}	R Documentation

Topp-Leone Distribution Family Function

Description

Estimating the parameter of the Topp-Leone distribution by maximum likelihood estimation.

Usage

topple(lshape = "logit", zero = NULL, gshape = ppoints(8))

Arguments

lshape, zero, gshape
More information is at CommonVGAMffArguments.

Details

The Topple distribution has a probability density function that can be written

f(y;s) = 2 * s * (1 - y) * (y * (2-y))^(s-1)

for 0<y<1 and shape parameter 0<s<1. The mean of Y is 1 - 4^s [Gamma(1+s)]^2 / Gamma(2 + 2s) (returned as the fitted values).

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Note

Fisher-scoring and Newton-Raphson are the same here. A related distribution is the triangle distribution. This VGAM family function handles multiple responses.

Author(s)

T. W. Yee

References

Topp, C. W. and F. C. Leone (1955) A family of J-shaped frequency functions. Journal of the American Statistical Association, 50, 209–219.

See Also

Topple, Triangle.

Examples

tdata <- data.frame(y = rtopple(1000, shape = logit(1, inverse = TRUE)))
tfit <- vglm(y ~ 1, topple, data = tdata, trace = TRUE, crit = "coef")
coef(tfit, matrix = TRUE)
Coef(tfit)

[Package VGAM version 1.0-6 Index]