toxop {VGAM}	R Documentation

Toxoplasmosis Data

Description

Toxoplasmosis data in 34 cities in El Salvador.

Usage

data(toxop)

Format

A data frame with 34 observations on the following 4 variables.

rainfall

a numeric vector; the amount of rainfall in each city.

ssize

a numeric vector; sample size.

cityNo

a numeric vector; the city number.

positive

a numeric vector; the number of subjects testing positive for the disease.

Details

See the references for details.

Source

See the references for details.

References

Efron, B. (1978) Regression and ANOVA With zero-one data: measures of residual variation. Journal of the American Statistical Association, 73, 113–121.

Efron, B. (1986) Double exponential families and their use in generalized linear regression. Journal of the American Statistical Association, 81, 709–721.

See Also

double.expbinomial.

Examples

## Not run:  with(toxop, plot(rainfall, positive / ssize, col = "blue"))
plot(toxop, col = "blue") 
## End(Not run)

[Package VGAM version 1.0-6 Index]