undocumented-methods {VGAM}	R Documentation

Undocumented Methods Functions

Description

Lots of undocumented methods functions are aliased here. In the VGAM package there are currently many objects/methods/classes which are currently internal and/or undocumented. The help file suppresses the warnings when the package is 'CHECK'ed.

Methods

There are many methods and these will be documented over time.

object

This argument is often used, and it is the primary object from which the function operates on.

[Package VGAM version 1.0-6 Index]