vplot.profile {VGAM}	R Documentation

Plotting Functions for 'profile' Objects

Description

plot and pairs methods for objects of class "profile", but renamed as vplot and vpairs.

Usage

vplot.profile(x, ...)
vpairs.profile(x, colours = 2:3, ...)

Arguments

x
an object inheriting from class "profile".

colours
Colours to be used for the mean curves conditional on x and y respectively.

...
arguments passed to or from other methods.

Details

See profile.glm for details.

Author(s)

T. W. Yee adapted this function from profile.glm, written originally by D. M. Bates and W. N. Venables. (For S in 1996.)

See Also

profilevglm, confintvglm, lrt.stat, profile.glm, profile.nls.

Examples

pneumo <- transform(pneumo, let = log(exposure.time))
fit1 <- vglm(cbind(normal, mild, severe) ~ let, acat,
             trace = TRUE, data = pneumo)
pfit1 <- profile(fit1, trace = FALSE)
## Not run: 
vplot.profile(pfit1)
vpairs.profile(pfit1)

## End(Not run)

[Package VGAM version 1.0-6 Index]