waitakere {VGAM}	R Documentation

Waitakere Ranges Data

Description

The waitakere data frame has 579 rows and 18 columns. Altitude is explanatory, and there are binary responses (presence/absence = 1/0 respectively) for 17 plant species.

Usage

data(waitakere)

Format

This data frame contains the following columns:

agaaus

Agathis australis, or Kauri

beitaw

Beilschmiedia tawa, or Tawa

corlae

Corynocarpus laevigatus

cyadea

Cyathea dealbata

cyamed

Cyathea medullaris

daccup

Dacrydium cupressinum

dacdac

Dacrycarpus dacrydioides

eladen

Elaecarpus dentatus

hedarb

Hedycarya arborea

hohpop

Species name unknown

kniexc

Knightia excelsa, or Rewarewa

kuneri

Kunzea ericoides

lepsco

Leptospermum scoparium

metrob

Metrosideros robusta

neslan

Nestegis lanceolata

rhosap

Rhopalostylis sapida

vitluc

Vitex lucens, or Puriri

altitude

meters above sea level

Details

These were collected from the Waitakere Ranges, a small forest in northern Auckland, New Zealand. At 579 sites in the forest, the presence/absence of 17 plant species was recorded, as well as the altitude. Each site was of area size 200m^2.

Source

Dr Neil Mitchell, University of Auckland.

See Also

hunua.

Examples

fit <- vgam(agaaus ~ s(altitude, df = 2), binomialff, waitakere)
head(predict(fit, waitakere, type = "response"))
## Not run:  plot(fit, se = TRUE, lcol = "orange", scol = "blue") 

[Package VGAM version 1.0-6 Index]