waldff {VGAM}	R Documentation

Wald Distribution Family Function

Description

Estimates the parameter of the standard Wald distribution by maximum likelihood estimation.

Usage

waldff(llambda = "loge", ilambda = NULL)

Arguments

llambda,ilambda
See CommonVGAMffArguments for information.

Details

The standard Wald distribution is a special case of the inverse Gaussian distribution with mu=1. It has a density that can be written as

f(y;mu,lambda) = sqrt(lambda/(2*pi*y^3)) * exp(-lambda*(y-1)^2/(2*y))

where y>0 and lambda>0. The mean of Y is 1 (returned as the fitted values) and its variance is 1/lambda. By default, eta=log(lambda).

Value

An object of class "vglmff" (see vglmff-class). The object is used by modelling functions such as vglm, and vgam.

Note

The VGAM family function inv.gaussianff estimates the location parameter mu too.

Author(s)

T. W. Yee

References

Johnson, N. L. and Kotz, S. and Balakrishnan, N. (1994) Continuous Univariate Distributions, 2nd edition, Volume 1, New York: Wiley.

See Also

inv.gaussianff, rinv.gaussian.

Examples

wdata <- data.frame(y = rinv.gaussian(n = 1000, mu =  1, lambda = exp(1)))
wfit <- vglm(y ~ 1, waldff(ilambda = 0.2), data = wdata, trace = TRUE)
coef(wfit, matrix = TRUE)
Coef(wfit)
summary(wfit)

[Package VGAM version 1.0-6 Index]