wrapup.smart {VGAM}	R Documentation

Cleans Up After Smart Prediction

Description

wrapup.smart deletes any variables used by smart prediction. Needed by both the modelling function and the prediction function.

Usage

wrapup.smart()

Details

The variables to be deleted are .smart.prediction, .smart.prediction.counter, and .smart.prediction.mode. The function wrapup.smart is useful in R because these variables are held in smartpredenv.

See Also

setup.smart.

Examples

## Not run: # Place this inside modelling functions such as lm, glm, vglm.
wrapup.smart()  # Put at the end of lm

## End(Not run)

[Package VGAM version 1.0-6 Index]