Yules {VGAM}	R Documentation

Yule-Simon Distribution

Description

Density, distribution function, quantile function and random generation for the Yule-Simon distribution.

Usage

dyules(x, shape, log = FALSE)
pyules(q, shape, lower.tail = TRUE, log.p = FALSE)
qyules(p, shape)
ryules(n, shape)

Arguments

x, q, p, n
Same meaning as in Normal.

shape
See yulesimon.

log, lower.tail, log.p
Same meaning as in pnorm or qnorm.

Details

See yulesimon, the VGAM family function for estimating the parameter, for the formula of the probability density function and other details.

Value

dyules gives the density, pyules gives the distribution function, qyules gives the quantile function, and ryules generates random deviates.

Note

Numerical problems may occur with qyules() when p is very close to 1.

Author(s)

T. W. Yee

See Also

yulesimon.

Examples

dyules(1:20, 2.1)
ryules(20, 2.1)

round(1000 * dyules(1:8, 2))
table(ryules(1000, 2))

## Not run:  x <- 0:6
plot(x, dyules(x, shape = 2.2), type = "h", las = 1, col = "blue") 
## End(Not run)

[Package VGAM version 1.0-6 Index]