zero {VGAM}	R Documentation

The zero Argument in VGAM Family Functions

Description

The zero argument allows users to conveniently model certain linear/additive predictors as intercept-only.

Details

Often a certain parameter needs to be modelled simply while other parameters in the model may be more complex, for example, the lambda parameter in LMS-Box-Cox quantile regression should be modelled more simply compared to its mu parameter. Another example is the xi parameter in a GEV distribution which is should be modelled simpler than its mu parameter. Using the zero argument allows this to be fitted conveniently without having to input all the constraint matrices explicitly.

The zero argument can be assigned an integer vector from the set {1:M} where M is the number of linear/additive predictors. Full details about constraint matrices can be found in the references. See CommonVGAMffArguments for more information.

Value

Nothing is returned. It is simply a convenient argument for constraining certain linear/additive predictors to be an intercept only.

Warning

The use of other arguments may conflict with the zero argument. For example, using constraints to input constraint matrices may conflict with the zero argument. Another example is the argument parallel. In general users should not assume any particular order of precedence when there is potential conflict of definition. Currently no checking for consistency is made.

The argument zero may be renamed in the future to something better.

Side Effects

The argument creates the appropriate constraint matrices internally.

Note

In all VGAM family functions zero = NULL means none of the linear/additive predictors are modelled as intercepts-only. Almost all VGAM family function have zero = NULL as the default, but there are some exceptions, e.g., binom2.or.

Typing something like coef(fit, matrix = TRUE) is a useful way to ensure that the zero argument has worked as expected.

Author(s)

T. W. Yee

References

Yee, T. W. and Wild, C. J. (1996) Vector generalized additive models. Journal of the Royal Statistical Society, Series B, Methodological, 58, 481–493.

Yee, T. W. and Hastie, T. J. (2003) Reduced-rank vector generalized linear models. Statistical Modelling, 3, 15–41.

See Also

CommonVGAMffArguments, constraints.

Examples

args(multinomial)
args(binom2.or)
args(gpd)

#LMS quantile regression example
fit <- vglm(BMI ~ sm.bs(age, df = 4), lms.bcg(zero = c(1, 3)),
            data = bmi.nz, trace = TRUE)
coef(fit, matrix = TRUE)

[Package VGAM version 1.0-6 Index]