Log {VGAM}	R Documentation

Logarithmic Distribution

Description

Density, distribution function, and random generation for the logarithmic distribution.

Usage

dlog(x, prob, log = FALSE)
plog(q, prob, log.p = FALSE)
rlog(n, prob, Smallno = 1.0e-6)

Arguments

x, q
Vector of quantiles. For the density, it should be a vector with positive integer values in order for the probabilities to be positive.

n
number of observations. Same as in runif.

prob
The parameter value c described in in logff. Here it is called prob because 0<c<1 is the range. For rlog() this parameter must be of length 1.

log, log.p
Logical. If log.p = TRUE then all probabilities p are given as log(p).

Smallno
Numeric, a small value used by the rejection method for determining the upper limit of the distribution. That is, plog(U, prob) > 1-Smallno where U is the upper limit.

Details

The details are given in logff.

Value

dlog gives the density, plog gives the distribution function, and rlog generates random deviates.

Note

Given some response data, the VGAM family function logff estimates the parameter prob. For plog(), if argument q contains large values and/or q is long in length then the memory requirements may be very high. Very large values in q are handled by an approximation by Owen (1965).

Author(s)

T. W. Yee

References

Forbes, C., Evans, M., Hastings, N. and Peacock, B. (2011) Statistical Distributions, Hoboken, NJ, USA: John Wiley and Sons, Fourth edition.

See Also

logff.

Examples

dlog(1:20, 0.5)
rlog(20, 0.5)

## Not run:  prob <- 0.8; x <- 1:10
plot(x, dlog(x, prob = prob), type = "h", ylim = 0:1,
     sub = "prob=0.8", las = 1, col = "blue", ylab = "Probability",
     main = "Logarithmic distribution: blue=density; orange=distribution function")
lines(x + 0.1, plog(x, prob = prob), col = "orange", lty = 3, type = "h") 
## End(Not run)

[Package VGAM version 1.0-2 Index]