summaryvglm {VGAM}	R Documentation

Summarizing Vector Generalized Linear Model Fits

Description

These functions are all methods for class vglm or summary.vglm objects.

Usage

summaryvglm(object, correlation = FALSE,
            dispersion = NULL, digits = NULL, presid = TRUE,
            signif.stars = getOption("show.signif.stars"),
            nopredictors = FALSE, ...)
## S3 method for class 'summary.vglm'
show(x, digits = max(3L, getOption("digits") - 3L),
           quote = TRUE, prefix = "", presid = TRUE,
           signif.stars = NULL, nopredictors = NULL, ...)

Arguments

object
an object of class "vglm", usually, a result of a call to vglm.

x
an object of class "summary.vglm", usually, a result of a call to summaryvglm().

dispersion
used mainly for GLMs. See summary.glm.

correlation
logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed.

digits
the number of significant digits to use when printing.

signif.stars
logical; if TRUE, ‘significance stars’ are printed for each coefficient.

presid
Pearson residuals; print out some summary statistics of these?

quote
Fed into print().

nopredictors
logical; if TRUE the names of the linear predictors are not printed out. The default is that they are.

prefix
Not used.

...
Not used.

Details

show.summary.vglm() tries to be smart about formatting the coefficients, standard errors, etc. and additionally gives ‘significance stars’ if signif.stars is TRUE. The coefficients component of the result gives the estimated coefficients and their estimated standard errors, together with their ratio. This third column is labelled z value regardless of whether the dispersion is estimated or known (or fixed by the family). A fourth column gives the two-tailed p-value corresponding to the z ratio based on a Normal reference distribution. In general, the t distribution is not used, but the normal distribution is used.

Correlations are printed to two decimal places (or symbolically): to see the actual correlations print summary(object)@correlation directly.

It is possible for programmers to write a methods function to print out extra quantities when summary(vglmObject) is called. The generic function is summaryvglmS4VGAM(), and one can use the S4 function setMethod to compute the quantities needed. Also needed is the generic function is showsummaryvglmS4VGAM() to actually print the quantities out.

Value

summaryvglm returns an object of class "summary.vglm"; see summary.vglm-class.

See Also

vglm, confintvglm, vcovvlm, summary.glm, summary.lm, summary.

Examples

## For examples see example(glm)
pneumo <- transform(pneumo, let = log(exposure.time))
(fit <- vglm(cbind(normal, mild, severe) ~ let, acat, data = pneumo))
coef(fit, matrix = TRUE)
summary(fit)
coef(summary(fit))

[Package VGAM version 1.0-2 Index]